% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_external_phenotype.R
\name{add_external_phenotype}
\alias{add_external_phenotype}
\title{A function to align genotype and phenotype data}
\usage{
add_external_phenotype(geno, geno_id = "sample.ID", pheno, pheno_id, pheno_col)
}
\arguments{
\item{geno}{An object created by \code{name_and_count_bigsnp()}}

\item{geno_id}{A character string indicating the ID column name in the 'fam'
element of the genotype data list. Defautls to 'sample.ID', equivalent to 'IID' in PLINK.}

\item{pheno}{A data frame with at least two columns: and ID column and a phenotype column}

\item{pheno_id}{A string specifying the name of the ID column in \code{pheno}}

\item{pheno_col}{A string specifying the name of the phenotype column in \code{pheno}. This column will be used as the default \code{y} argument to 'plmm()'.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    Nothing is returned; instead, the results are written to an RDS file. 
             By default, the RDS file in `geno` is overwritten. You may change this behavior by specifying a new 
             filename to `rdsfile` 
}\if{html}{\out{</div>}}
}
\description{
A function to align genotype and phenotype data
}
\keyword{internal}
