% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_famfile_ids.R
\name{align_famfile_ids}
\alias{align_famfile_ids}
\title{A helper function to support \code{process_plink()}}
\usage{
align_famfile_ids(id_var, quiet, add_predictor, og_plink_ids)
}
\arguments{
\item{id_var}{String specifying which variable in the PLINK file was the unique id: 'FID' or 'IID'}

\item{quiet}{Logical: should a message be printed?}

\item{add_predictor}{External data to include in design matrix. This is the add_predictors... arg in \code{process_plink()}}

\item{og_plink_ids}{Character vector with the PLINK ids (FID or IID) from the \emph{original} data (i.e., the data before any subsetting from handling missing phenotypes)}
}
\value{
An object of the same type as add_predictor
}
\description{
A helper function to support \code{process_plink()}
}
\keyword{internal}
