% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_std.R
\name{big_std}
\alias{big_std}
\title{big_std
A function to standardize a matrix stored as an FBM object}
\usage{
big_std(X, center = NULL, scale, ns = NULL, std_bk_extension = NULL)
}
\arguments{
\item{X}{An file-backed design matrix}

\item{center}{A vector of centering values}

\item{scale}{A vector of scaling values}

\item{ns}{A vector with the indices marking the Non-Singular columns of X. Defaults to NULL (meaning all columns are assumed to be nonsingular).}

\item{std_bk_extension}{The string specifying the backingfile for the standardized copy of X. Default uses a temporary directory.}
}
\value{
A standardized matrix of class FBM
}
\description{
big_std
A function to standardize a matrix stored as an FBM object
}
\keyword{internal}
