% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm2bm.R
\name{bm2fbm}
\alias{bm2fbm}
\title{Create an FBM that points to the same backingfile as a big.matrix
Credit to Florian Prive for this function}
\usage{
bm2fbm(bm)
}
\arguments{
\item{bm}{A \code{big.matrix} object}
}
\value{
An \code{FBM} object pointing to the same backingfile
}
\description{
Create an FBM that points to the same backingfile as a big.matrix
Credit to Florian Prive for this function
}
\details{
For an extended example, see the tutorial on the \code{bigstatsr} package site:
https://privefl.github.io/bigstatsr/articles/bigstatsr-and-bigmemory.html
}
\keyword{internal}
