% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_file_extension.R
\name{check_for_file_extension}
\alias{check_for_file_extension}
\title{check_for_file_extension: a function to make our package 'smart' enough to
handle .rds file extensions}
\usage{
check_for_file_extension(path)
}
\arguments{
\item{path}{A string specifying a file path that ends in a file name, e.g. "~/dir/my_file.rds"}
}
\value{
a string with a filepath \emph{without} an extension, e.g. "~/dir/my_file"
}
\description{
check_for_file_extension: a function to make our package 'smart' enough to
handle .rds file extensions
}
\keyword{internal}
