% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_k.R
\name{choose_k}
\alias{choose_k}
\title{a function to choose k, the number of eigenvalues to use in truncated SVD}
\usage{
choose_k(
  X,
  start = NULL,
  step = NULL,
  eps = 2,
  trace = TRUE,
  type = "F",
  returnKapprox = FALSE,
  returnK = FALSE
)
}
\arguments{
\item{X}{can be either:
\itemize{
\item the fully-imputed design matrix, or
\item a string specifying the path to an .rds object created by process_plink
}}

\item{start}{The starting number of eigenvalues. Defaults to floor(nrow(X)/3)}

\item{step}{The size (with respect to the number of observations) of the increments of increase in choosing k. Defaults to floor(nrow(X)/10).}

\item{eps}{The largest permissible difference between the true and approximate relatedness matrices. Defaults to 2.}

\item{trace}{Logical: should progress bars and messages be printed to the console? Defaults to TRUE.}

\item{type}{The type of norm to use in determining distance between true and approximate K. Defaults to 'F', for Frobenious norm. See Matrix::norm() for details.}

\item{returnKapprox}{Logical: in addition to the list of SVD components for approximated K, should the approximation be returned as a matrix? Defaults to FALSE.}

\item{returnK}{Logical: should the true K (as in relatedness_mat(X)) be returned? Defaults to FALSE.}
}
\value{
A list with at least 3 items:
\itemize{
\item K_svd: a list with the SVD components (d and u) for the approximated relatedness matrix K that can be passed to \code{plmm()}
\item k, the chosen number of eigenvalues
\item delt, the distance between the true and approximated K matrices
\item optional: if returnKapprox, K_approx will be returned
\item optional: if returnK, K will be returned
}
}
\description{
a function to choose k, the number of eigenvalues to use in truncated SVD
}
\examples{
\dontrun{
choose_k(X = admix$X, start = 10, step = 10, trace = TRUE)
}
 
}
