% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-cv-plmm.R
\name{coef.cv.plmm}
\alias{coef.cv.plmm}
\title{Coef method for "cv.plmm" class}
\usage{
\method{coef}{cv.plmm}(object, lambda, which = object$min, ...)
}
\arguments{
\item{object}{An object of class "cv.plmm."}

\item{lambda}{A numeric vector of lambda values.}

\item{which}{Vector of lambda indices for which coefficients to return. Defaults to lambda index with minimum CVE.}

\item{...}{Additional arguments (not used).}
}
\value{
Returns a named numeric vector. Values are the coefficients of the
model at the specified value of either \code{lambda} or \code{which}. Names are the
values of \code{lambda}.
}
\description{
Coef method for "cv.plmm" class
}
\examples{
\donttest{
#' cv_fit <- cv.plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X))
head(coef.cv.plmm(cv_fit))
}


}
