% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-list.R
\name{coef.list}
\alias{coef.list}
\title{Coef method for a list used in cross-validation (e.g. the \code{fit.i} within \code{cvf})}
\usage{
\method{coef}{list}(fit, lambda, which, drop, ...)
}
\arguments{
\item{fit}{A list with the components returned by \code{plmm_fit}}

\item{lambda}{A numeric vector of lambda values.}

\item{which}{Vector of lambda indices for which coefficients to return.}

\item{drop}{Logical.}

\item{...}{Additional arguments.}
}
\description{
Coef method for a list used in cross-validation (e.g. the \code{fit.i} within \code{cvf})
}
\keyword{internal}
