% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-lmm.R
\name{coef.lmm}
\alias{coef.lmm}
\title{Coef method for "lmm" class}
\usage{
\method{coef}{lmm}(object, which = 1:length(object$beta_vals), ...)
}
\arguments{
\item{object}{An object of class "lmm."}

\item{which}{Vector of indices for which coefficients to return.}

\item{...}{Additional arguments.}
}
\description{
Coef method for "lmm" class
}
\examples{
\dontrun{
fit <- lmm(X = pedigree$X, y = pedigree$clinical$y, K = pedigree$K)
coef.lmm(fit)
}

}
