% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-plmm.R
\name{coef.plmm}
\alias{coef.plmm}
\title{Coef method for \code{plmm} class}
\usage{
\method{coef}{plmm}(object, lambda, which = seq_along(object$lambda), drop = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{plmm}.}

\item{lambda}{A numeric vector of lambda values.}

\item{which}{Vector of lambda indices for which to return coefficients.}

\item{drop}{Logical. Should returned object be coerced to a vector if possible?}

\item{...}{Additional arguments.}
}
\value{
Either a numeric matrix (if model was fit on data stored in memory)
or a sparse matrix (if model was fit on data stored filebacked). Rownames are
feature names, columns are values of \code{lambda}.
}
\description{
Coef method for \code{plmm} class
}
\examples{
admix_design <- create_design(X = admix$X, y = admix$y)
fit <- plmm(design = admix_design)
coef(fit)[1:10, 41:45]
}
