% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_snpmatrix.R
\name{coerce_snpmatrix}
\alias{coerce_snpmatrix}
\title{Coerce SnpMatrix: a function to convert a SnpMatrix object into a matrix of numeric values}
\usage{
coerce_snpmatrix(snpmatrix)
}
\arguments{
\item{snpmatrix}{An object of class "SnpMatrix", as defined in the \code{snpStats} package.}
}
\value{
A matrix of numeric values
}
\description{
Coerce SnpMatrix: a function to convert a SnpMatrix object into a matrix of numeric values
}
\examples{
\dontrun{
cad <- process_plink(prefix = "cad", dataDir = plink_example(path="cad.fam", parent=T))
X <- coerce_snpmatrix(cad$genotypes)
}
}
