% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_variances.R
\name{compare_variances}
\alias{compare_variances}
\title{quick function to do some further simulations}
\usage{
compare_variances(nrep, K, sig_s, sig_eps, intercept = TRUE, center_y = FALSE)
}
\arguments{
\item{nrep}{Number of reps for simulation}

\item{K}{Realized relatedness matrix}

\item{sig_s}{variance due to structure}

\item{sig_eps}{variance due to noise}

\item{intercept}{Logical: defaults to TRUE.}

\item{center_y}{Logical: defaults to FALSE.}
}
\value{
list with two named vectors:
\itemize{
\item intercept: estimated eta values from the method that defines null model as intercept-only model
\item zero_mean: estimated eta values from the model that defines y to have mean 0 on rotated scale.
}
}
\description{
quick function to do some further simulations
}
