% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_blup.R
\name{compute_blup}
\alias{compute_blup}
\title{A function to compute the BLUP}
\usage{
compute_blup(fit, Xb, Sigma_21, idx)
}
\arguments{
\item{fit}{An object returned by \code{plmm()}}

\item{Xb}{Linear predictor}

\item{Sigma_21}{Covariance matrix between the training and the testing data. Extracted from \code{estimated_Sigma} that is generated using all observations}

\item{idx}{Vector of indices of the penalty parameter \code{lambda} at which predictions are required. By default, all indices are returned.}
}
\value{
A matrix of the linear predictors + the estimated random effects
}
\description{
A function to compute the BLUP
}
\keyword{internal}
