% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_variance.R
\name{construct_variance}
\alias{construct_variance}
\title{a function to create the estimated variance matrix from a PLMM fit}
\usage{
construct_variance(fit, K = NULL, eta = NULL)
}
\arguments{
\item{fit}{An object returned by \code{plmm()}}

\item{K}{An optional matrix}

\item{eta}{An optional numeric value between 0 and 1; if \code{fit} is not supplied, then this option must be specified.}
}
\value{
Sigma_hat, a matrix representing the estimated variance
}
\description{
a function to create the estimated variance matrix from a PLMM fit
}
