% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexMin.R
\name{convexMin}
\alias{convexMin}
\title{Calculate index for which objective function ceases to be locally convex}
\usage{
convexMin(
  b,
  X,
  penalty,
  gamma,
  l2,
  family = "gaussian",
  penalty.factor,
  a,
  Delta = NULL
)
}
\arguments{
\item{b}{Matrix of coefficient values.}

\item{X}{Design matrix.}

\item{penalty}{The penalty applied to the model. Either "MCP", "SCAD", or "lasso".}

\item{gamma}{The tuning parameter of the MCP/SCAD penalty. Default is 3 for MCP and 3.7 for SCAD.}

\item{l2}{L2.}

\item{family}{Only "gaussian" currently supported.}

\item{penalty.factor}{A multiplicative factor for the penalty applied to each coefficient. If supplied, penalty.factor must be a numeric vector of length equal to the number of columns of X. The purpose of penalty.factor is to apply differential penalization if some coefficients are thought to be more likely than others to be in the model. In particular, penalty.factor can be 0, in which case the coefficient is always in the model without shrinkage.}

\item{a}{Vector of intercept values.}

\item{Delta}{NULL.}
}
\description{
Calculate index for which objective function ceases to be locally convex
}
