% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_constant_features.R
\name{count_constant_features}
\alias{count_constant_features}
\title{A helper function to count constant features}
\usage{
count_constant_features(
  fbm,
  ind.row = bigstatsr::rows_along(fbm),
  outfile,
  quiet
)
}
\arguments{
\item{fbm}{An FBM object, usually the 'genotypes' component from a bigSNP list}

\item{ind.row}{Optional indicators for rows in fbm. Default includes all rows.}

\item{outfile}{String specifying name of log file}

\item{quiet}{Logical: should a message be printed to the console}
}
\value{
ns A numeric vector with the indicesof the non-singular columns of the matrix associated with \code{counts}
}
\description{
A helper function to count constant features
}
\keyword{internal}
