% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cores.R
\name{count_cores}
\alias{count_cores}
\title{A helper function to count the number of cores available on the current machine}
\usage{
count_cores()
}
\value{
A number of cores to use; if \code{parallel} is installed, this will be \code{parallel::detectCores()}. Otherwise, this returns a 1.
}
\description{
A helper function to count the number of cores available on the current machine
}
\keyword{internal}
