% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_in_memory.R
\name{create_design_in_memory}
\alias{create_design_in_memory}
\title{A function to create a design with an in-memory X matrix}
\usage{
create_design_in_memory(X, y, unpen = NULL)
}
\arguments{
\item{X}{A numeric matrix in which rows correspond to observations (e.g., samples) and columns correspond to features.}

\item{y}{A numeric vector representing the outcome for the model.
\strong{Note}: it is the responsibility of the user to ensure that the outcome_col and X have the same row order!}

\item{unpen}{An optional character vector with the names of columns to mark as unpenalized (i.e., these features would always be included in a model).
\strong{Note}: if you choose to use this option, X must have column names.}
}
\value{
A list with elements including a standardized X and model design information
}
\description{
A function to create a design with an in-memory X matrix
}
\keyword{internal}
