% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_log.R
\name{create_log}
\alias{create_log}
\title{Create the \code{.log} file}
\usage{
create_log(outfile)
}
\arguments{
\item{outfile}{String specifying the name of the to-be-created file, \emph{without} extension}
}
\value{
Nothing is returned, instead a text file with the suffix \code{.log} is created.
}
\description{
Create the \code{.log} file
}
\keyword{internal}
