% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvf.R
\name{cvf}
\alias{cvf}
\title{Cross-validation internal function for \code{cv_plmm()}}
\usage{
cvf(i, fold, type, cv_args, ...)
}
\arguments{
\item{i}{Fold number to be excluded from fit.}

\item{fold}{n-length vector of fold-assignments.}

\item{type}{A character argument indicating what should be returned from \code{predict.plmm()}. If \code{type = 'lp'} predictions are based on the linear predictor, \eqn{X \beta}.
If \code{type = 'individual'} predictions are based on the linear predictor plus the estimated random effect (BLUP).}

\item{cv_args}{List of additional arguments to be passed to plmm.}

\item{...}{Optional arguments to \code{predict_within_cv()}}
}
\value{
A list with three elements:
\itemize{
\item \code{loss}: a numeric vector with the loss at each value of lambda
\item \code{nl}: a numeric value indicating the number of lambda values used
\item \code{yhat}: a numeric value with the predicted outcome values at each lambda
}
}
\description{
Internal function for \code{cv_plmm()} which calls \code{plmm()} on a fold subset of the original data.
}
\keyword{internal}
