% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvf_predict_unrotated.R
\name{cvf_predict_unrotated}
\alias{cvf_predict_unrotated}
\title{Cross-validation internal function for cv.plmm that computes prediction based on the unrotated data.}
\usage{
cvf_predict_unrotated(i, XX, y, X_unrotated, y_unrotated, fold, cv.args)
}
\arguments{
\item{i}{Fold number to be excluded from fit.}

\item{XX}{Transformed design matrix including intercept. May include clinical covariates and other non-SNP data. If this is the case, X_for_K should be supplied with a  matrix containing only SNP data for computation of GRM.}

\item{y}{Continuous outcome vector.}

\item{X_unrotated}{Non-transformed design matrix, which should not include an intercept. May include clinical covariates and other non-SNP data.}

\item{y_unrotated}{Non-transformed outcome vector.}

\item{fold}{n-length vector of fold-assignments.}

\item{cv.args}{List of additional arguments to be passed to plmm.}
}
\description{
Internal function for cv.plmm which calls plmm on a fold subset of the original data.
}
