% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{delta}
\alias{delta}
\title{a helper function to calculate the 'distance' between 2 matrices}
\usage{
delta(A, B, type = "F")
}
\arguments{
\item{A}{The matrix to be compared with B}

\item{B}{The second matrix (I think of this as the approximation to A)}

\item{type}{String indicating the type of norm to be used. See Matrix::norm() for details}
}
\description{
a helper function to calculate the 'distance' between 2 matrices
}
\keyword{internal}
