% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_K.R
\name{eigen_K}
\alias{eigen_K}
\title{A function to take the eigendecomposition of K}
\usage{
eigen_K(std_X)
}
\arguments{
\item{std_X}{The \emph{standardized} design matrix.}
}
\value{
A list with three elements:
\itemize{
\item \code{s}: The non-zero eigenvalues of K
\item \code{U}: The eigenvectors of K associated with s
\item \code{K}: The fully computed K matrix
}
}
\description{
Note: This is faster than taking SVD of X when p >> n
}
\keyword{internal}
