% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm2bm.R
\name{fbm2bm}
\alias{fbm2bm}
\title{Functions to convert between FBM and big.matrix type objects}
\usage{
fbm2bm(fbm, desc = FALSE)
}
\arguments{
\item{fbm}{An FBM object; see \code{bigstatsr::FBM()} for details}

\item{desc}{Logical: is the descriptor file desired (as opposed to the filebacked big matrix)? Defaults to FALSE.}
}
\value{
a \code{big.matrix} - see \code{bigmemory::filebacked.big.matrix()} for details
}
\description{
Functions to convert between FBM and big.matrix type objects
}
\keyword{internal}
