% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_sans_ext.R
\name{file_sans_ext}
\alias{file_sans_ext}
\title{helper function to get the file path of a file \emph{without} the extension}
\usage{
file_sans_ext(path)
}
\arguments{
\item{path}{The path of a file}
}
\value{
path_sans_ext The filepath without the extension
}
\description{
helper function to get the file path of a file \emph{without} the extension
}
\keyword{internal}
