% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_signs.R
\name{flip_signs}
\alias{flip_signs}
\title{a function to check for flipped signs of an SVD}
\usage{
flip_signs(X, U, d, V, svd_list = NULL)
}
\arguments{
\item{X}{matrix whose SVD is UDV^t. SVD may be truncated.}

\item{U}{matrix of left singular vectors}

\item{d}{vector of singular values}

\item{V}{matrix of right singular vectors}

\item{svd_list}{optional list with components u, d, and v}
}
\description{
a function to check for flipped signs of an SVD
}
\details{
Algorithm here is inspired by Bro2007: https://www.osti.gov/servlets/purl/920802
}
\examples{
\dontrun{
svd_X <- svd(admix$X)
svd_X$u[1:3, 1:3]
flip_signs(X = admix$X, svd_list = svd_X) -> foo_mat
foo_mat$U[1:3, 1:3]

K <- relatedness_mat(admix$X)
K[1:4, 1:4]
full_svd <- svd(K)
full_svd$u[1:4, 1:4]
trunc_svd <- RSpectra::svds(K, k = 10)
trunc_svd$u[1:4, 1:4] # notice: signs flip (which is a 'bad sign')

# let's fix this: 
corrected_trunc_svd <- flip_signs(X = K, svd_list = trunc_svd)
corrected_trunc_svd$U[1:4, 1:4] 
}
}
\keyword{internal}
