% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_K.R
\name{generate_K}
\alias{generate_K}
\title{A function to generate K matrices for simulation}
\usage{
generate_K(n_per_group, n_groups = 10, mu = 0.7)
}
\arguments{
\item{n_per_group}{Number of observations per group (e.g., per family, per batch, ...)}

\item{n_groups}{Number of groups}

\item{mu}{Optional vector of length n_groups with the covariance value for each group. Defaults to 0.7 for all groups.}

\item{plot}{Logical: should a plot of K be shown? Defaults to TRUE.}
}
\value{
A K matrix
}
\description{
A function to generate K matrices for simulation
}
\details{
For internal use in running tests
}
\keyword{internal}
