% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genGammaUnscaled.R
\name{genGammaUnscaled}
\alias{genGammaUnscaled}
\title{Generate a vector of environmental confounding effects}
\usage{
genGammaUnscaled(structureGamma, J)
}
\arguments{
\item{structureGamma}{A character argument describing the desired structure.}

\item{J}{Number of subpopulations.}
}
\description{
This function allows you to simulate an environmental confounding effect. This is used by \code{genDataPS()}, which scales the generated vector in order to comply with the user specified desired variance in the outcome attributable to the environmental effect.
}
\examples{
genGammaUnscaled(structureGamma = "linear_concordant", J = 3)
}
