% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRelatednessMat.R
\name{genRelatednessMat}
\alias{genRelatednessMat}
\title{Generate a relatedness matrix}
\usage{
genRelatednessMat(X, ...)
}
\arguments{
\item{X}{Either a numeric matrix of genotypes (subjects in rows, SNPs in columns), or a character prefix for PLINK bed/bim/fam files.}

\item{...}{Optional arguments in the case that a character prefix is specified.}
}
\description{
This function allows you to simulate an environmental confounding effect. This is used by \code{genDataPS()}, which scales the generated vector in order to comply with the user specified desired variance in the outcome attributable to the environmental effect.
}
