% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genXps.R
\name{genXps}
\alias{genXps}
\title{Generate an X matrix with population structure}
\usage{
genXps(
  n,
  nJ,
  p,
  structureX = c("admixture", "indep_subpops", "1d_linear", "1d_circular",
    "independent", "other"),
  Fst = NULL,
  inbr = c("homogeneous", "heterogeneous"),
  standardizeX = TRUE,
  plot = FALSE,
  structureX_other = NULL,
  sampleCols = TRUE
)
}
\arguments{
\item{n}{Number of observations/samples to simulate.}

\item{nJ}{Number of observations in each subpopulation. The length of nJ corresponds to the number of subpopulations.}

\item{p}{Number of SNPs to simulate.}

\item{structureX}{Type of structure to simulate.}

\item{Fst}{The desired final FST of the admixed individuals. Ranges from 0 to 1. A high Fst implies greater differentiation among populations. Defaults to 0.1 if structureX = 1d_linear and 0.2 if structureX = indep_subpops. Otherwise defaults to NULL.}

\item{inbr}{Indicates whether the desired inbreeding is homogeneous or heterogeneous. Defaults to heterogeneous.}

\item{standardizeX}{Should the generated X matrix be standardized? Defaults to TRUE.}

\item{plot}{Should a plot of the kinship matrix be generated? Defaults to FALSE.}

\item{structureX_other}{If \code{structureX == "other"}, an matrix or SnpMatrix object with subjects in rows and SNPs in columns to be used to generate pseudophenotypes must be supplied here.}

\item{sampleCols}{A logical flag for whether the columns of the resultant X matrix should be scrambled. This may be desirable if the causal SNPs should change from one simulation to the next. Defaults to TRUE.}
}
\description{
This function allows you to simulate structured genetic data (SNP).
}
\examples{
sim_X <- genXps(n = 200, nJ = rep(100, 2), p = 1000, structureX = "indep_subpops", inbr = "heterogeneous")
}
