% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hostname.R
\name{get_hostname}
\alias{get_hostname}
\title{a function to return the computer's host name}
\usage{
get_hostname()
}
\value{
String with hostname of current machine
}
\description{
a function to return the computer's host name
}
\keyword{internal}
