% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gic.R
\name{gic}
\alias{gic}
\title{General information criterion method of selecting lambda for "plmm" class}
\usage{
gic(fit, ic = c("bic", "hdbic"), SUX, SUy, S, eta)
}
\arguments{
\item{fit}{An object of class "plmm."}

\item{ic}{Information criterion that should be used to select lambda. Currently supports BIC or HDBIC. Defaults to BIC.}

\item{SUX}{Rotated design matrix including rotated intercept and unpenalized columns, if present. If not returned as part of \code{plmm} because \code{returnX == FALSE}, must be supplied explicitly.}

\item{SUy}{Rotated outcome vector. If not returned as part of \code{plmm} because \code{returnX == FALSE}, must be supplied explicitly.}

\item{S}{Eigenvalues from similarity matrix used for model fitting. If not returned as part of \code{plmm} because \code{returnX == FALSE}, must be supplied explicitly.}

\item{eta}{Estimated $eta$ value from object fit.}
}
\description{
General information criterion method of selecting lambda for "plmm" class
}
