% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_data.R
\name{impute_data}
\alias{impute_data}
\title{A function to implement simple imputation on a filebacked matrix}
\usage{
impute_data(X, rds_dir, prefix, na_idx, impute_method)
}
\arguments{
\item{X}{A filebacked matrix of data - must be numeric}

\item{rds_dir}{The path to the directory in which you want to create
the new '.rds' and '.bk' files. Defaults to \code{data_dir}}

\item{prefix}{The prefix of the '.bk' and '.rds' files associated with \code{X}}

\item{na_idx}{Indices for columns with missing values}

\item{impute_method}{Three options here:
\itemize{
\item 'mean' = fill in missing values with the mean
\item 'median' = fill in missing values with the median
\item 'mode' = fill in missing values with a mode
Note: 'the' mode may not exist; x could have multiple modes
}}
}
\value{
A filebacked matrix of imputed data
}
\description{
A function to implement simple imputation on a filebacked matrix
}
\keyword{internal}
