% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lam_names.R
\name{lam_names}
\alias{lam_names}
\title{Generate nicely formatted lambda vec}
\usage{
lam_names(l)
}
\arguments{
\item{l}{Vector of lambda values.}
}
\value{
A character vector of formatted lambda value names
}
\description{
Generate nicely formatted lambda vec
}
\keyword{internal}
