% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamNames.R
\name{lamNames}
\alias{lamNames}
\title{Generate nicely formatted lambda vec}
\usage{
lamNames(l)
}
\arguments{
\item{l}{Vector of lambda values.}
}
\value{
A character vector of formatted lambda value names
}
\description{
Generate nicely formatted lambda vec
}
\examples{
l <- c(1, 0.1, 0.01, 0.001) 
lamNames(l)
}
\keyword{internal}
