% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_pca10.R
\name{lasso_pca10}
\alias{lasso_pca10}
\title{Fit a linear model with lasso regularization and PC adjutment}
\usage{
lasso_pca10(X, y, p1, standardize = FALSE)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Continuous outcome vector.}

\item{p1}{Number of causal SNPs. Lambda will be selected such that <= p1 variables enter the model.}

\item{standardize}{Should standardization be performed within \code{glmnet()}? Defaults to FALSE.}
}
\description{
This function allows you to fit a linear model via lasso-penalized maximum likelihood, adjusted for the first 10 principal components (unpenalized), with the same output values as those of lmm_lasso_eta. Primarily used for simulation and comparison with lmm_lasso_eta.
}
