% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lippert.R
\name{lippert_estimate_eta}
\alias{lippert_estimate_eta}
\title{the estimation of eta using the Lippert (2011) derivation}
\usage{
lippert_estimate_eta(n, s, U, y)
}
\arguments{
\item{n}{Number of observations}

\item{s}{Eigenvalues of K}

\item{U}{Eigenvectors of K}

\item{y}{Outcome vector}
}
\description{
the estimation of eta using the Lippert (2011) derivation
}
