% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lippert.R
\name{lippert_test_eta_estimation}
\alias{lippert_test_eta_estimation}
\title{a function to write tests of estimate_eta}
\usage{
lippert_test_eta_estimation(
  sig_s,
  sig_eps,
  K,
  intercept = TRUE,
  center_y = FALSE,
  y_skew = 0,
  return_y = FALSE,
  ...
)
}
\arguments{
\item{sig_s}{Variance attributable to structure}

\item{sig_eps}{Variance of random error}

\item{K}{Matrix to use as relatedness matrix.}

\item{intercept}{Logical: should the outcome be simulated with an intercept? Defaults to TRUE.}

\item{center_y}{Logical: should y be centered? Defaults to FALSE. (Note: this option only makes sense if intercept = TRUE)}

\item{y_skew}{Numeric value: if nonzero, this will control how much skewness will be added to the noise. Defaults to 0.}

\item{return_y}{Logical: should simulated outcome values be returned? Defaults to FALSE.}

\item{...}{Additional args to pass into \code{estimate_eta()}}
}
\description{
a function to write tests of estimate_eta
}
\keyword{internal}
