% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm}
\alias{lmm}
\title{a function to implement a mixed model for structured data \emph{without} penalization}
\usage{
lmm(
  X,
  y,
  k = NULL,
  K = NULL,
  eta_star = NULL,
  eps = 1e-04,
  max.iter = 10000,
  dfmax = ncol(X) + 1,
  warn = TRUE,
  init = rep(0, ncol(X)),
  returnX = TRUE,
  trace = FALSE
)
}
\arguments{
\item{X}{Design matrix. MUST NOT be high dimensional -- the modeling here will call \code{lm()}. For high-dimensional data, use \code{plmm()} instead.}

\item{y}{Continuous outcome vector.}

\item{k}{An integer specifying the number of singular values to be used in the approximation of the rotated design matrix. This argument is passed to \code{RSpectra::svds()}. Defaults to \code{min(n, p) - 1}, where n and p are the dimensions of the \emph{standardized} design matrix.}

\item{K}{Similarity matrix used to rotate the data. This should be (1) a known matrix that reflects the covariance of y, (2) an estimate (Default is \eqn{\frac{1}{p}(XX^T)}), or (3) a list as created by choose_k.}

\item{eta_star}{Optional argument to input a specific eta term rather than estimate it from the data. If K is a known covariance matrix that is full rank, this should be 1.}

\item{eps}{Convergence threshold. The algorithm iterates until the RMSD for the change in linear predictors for each coefficient is less than eps. Default is \code{1e-4}.}

\item{max.iter}{Maximum number of iterations (total across entire path). Default is 10000.}

\item{dfmax}{Upper bound for the number of nonzero coefficients. Default is no upper bound. However, for large data sets, computational burden may be heavy for models with a large number of nonzero coefficients.}

\item{warn}{Return warning messages for failures to converge and model saturation? Default is TRUE.}

\item{init}{Initial values for coefficients. Default is 0 for all columns of X.}

\item{returnX}{Return the standardized design matrix along with the fit? By default, this option is turned on if X is under 100 MB, but turned off for larger matrices to preserve memory.}

\item{trace}{If set to TRUE, inform the user of progress by announcing the beginning of each step of the modeling process. Default is FALSE.}
}
\value{
A list including the estimated coefficients on the original scale, as well as other model fitting details
}
\description{
NB: this function is simply a wrapper for lmm_prep -> lmm_fit -> lmm_format
}
\examples{
fit <- lmm(X = pedigree$X, y = pedigree$clinical$y, K = pedigree$K)
fit$beta_vals
}
