% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_format.R
\name{lmm_format}
\alias{lmm_format}
\title{LMM format: a function to format the output of a model constructed with \code{lmm_fit}}
\usage{
lmm_format(fit, dfmax = fit$ncol_X + 1, X)
}
\arguments{
\item{fit}{A list of parameters describing the output of a model constructed with \code{lmm_fit}}

\item{dfmax}{dfmax Upper bound for the number of nonzero coefficients. Default is no upper bound. However, for large data sets, computational burden may be heavy for models with a large number of nonzero coefficients.}

\item{X}{Design matrix. May include clinical covariates and other non-SNP data.}
}
\value{
A list with the components:
\itemize{
\item beta_vals: The estimated beta values at each value of lambda
\item eta: The estimated eta value
\item ns_idx: COME BACK HERE
\item iter: The number of iterations at each value of lambda (MAYBE take this out)
\item converged: The convergence status at each value of lambda
\item X: if returnX = TRUE and size SUX < 100 Mb, the original X will be returned
}
}
\description{
LMM format: a function to format the output of a model constructed with \code{lmm_fit}
}
\keyword{internal}
