% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_ggmix.R
\name{lmm_ggmix}
\alias{lmm_ggmix}
\title{Fit a linear mixed model with lasso regularization}
\usage{
lmm_ggmix(X, y, p1, standardize = FALSE, K = NULL)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Continuous outcome vector.}

\item{p1}{Number of causal SNPs. Lambda will be selected such that <= p1 variables enter the model.}

\item{standardize}{Should standardization be performed within \code{glmnet()}? Defaults to FALSE.}

\item{K}{Matrix used to compute the similarity matrix, K. For multi-chromosome analysis this may be supplied in order to perform a leave-one-chromosome-out correction. The objective here is to adjust for population stratification and unobserved confounding without rotating out the causal SNP effects. Default is \eqn{\frac{1}{p} XX^T}}
}
\description{
This function allows you to fit a linear mixed model via lasso-penalized maximum likelihood.
}
\examples{
lmm_ggmix_fit <- lmm_ggmix(X = admix$X, y = admix$y, p1 = 10, K = relatedness_mat(admix$X))
}
