% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_lasso_eta.R
\name{lmm_lasso_eta}
\alias{lmm_lasso_eta}
\title{Fit an LMM with lasso regularization}
\usage{
lmm_lasso_eta(X, y, p1, standardize = FALSE)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Continuous outcome vector.}

\item{p1}{Number of causal SNPs. Lambda will be selected such that <= p1 variables enter the model.}

\item{standardize}{Should standardization be performed within glmnet()? Defaults to FALSE.}
}
\description{
This function allows you to fit an approximate LMM (using the rotation method) with lasso regularization.
}
