% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_lasso_eta_null.R
\name{lmm_lasso_eta_null}
\alias{lmm_lasso_eta_null}
\title{Estimate eta using a null LmmLasso model}
\usage{
lmm_lasso_eta_null(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Continuous outcome vector.}
}
\description{
This function allows you to estimate eta (the signal to noise ratio, or narrow-sense variability) under the assumption of a null model in order to estimate the variance parameter, eta.
}
