% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_prep.R
\name{lmm_prep}
\alias{lmm_prep}
\title{a function to prepare data for an \emph{unpenalized} LMM}
\usage{
lmm_prep(
  X,
  y,
  k = NULL,
  K = NULL,
  diag_K = NULL,
  eta_star = NULL,
  returnX = TRUE,
  trace = FALSE,
  ...
)
}
\arguments{
\item{X}{Design matrix. May include clinical covariates and other non-SNP data.}

\item{y}{Continuous outcome vector.}

\item{k}{An integer specifying the number of singular values to be used in the approximation of the rotated design matrix. This argument is passed to \code{RSpectra::svds()}. Defaults to \code{min(n, p) - 1}, where n and p are the dimensions of the \emph{standardized} design matrix.}

\item{K}{Similarity matrix used to rotate the data. This should either be a known matrix that reflects the covariance of y, or an estimate (Default is \eqn{\frac{1}{p}(XX^T)}, where X is standardized).}

\item{diag_K}{Logical: should K be a diagonal matrix? This would reflect observations that are unrelated, or that can be treated as unrelated. Passed from \code{lmm()}.}

\item{eta_star}{Optional argument to input a specific eta term rather than estimate it from the data. If K is a known covariance matrix that is full rank, this should be 1.}

\item{returnX}{Return the standardized design matrix along with the fit? By default, this option is turned on if X is under 100 MB, but turned off for larger matrices to preserve memory.}

\item{trace}{If set to TRUE, inform the user of progress by announcing the beginning of each step of the modeling process. Default is FALSE.}

\item{...}{Not used yet}
}
\value{
List with these components:
\itemize{
\item ncol_X: The number of columns in the original design matrix
\item std_X: standardized design matrix
\item y: The vector of outcomes
\item S: The singular values of K
\item U: the left singular values of K (same as left singular values of X).
\item ns: the indices for the nonsingular values of std_X
\item snp_names: Formatted column names of the design matrix
}
}
\description{
a function to prepare data for an \emph{unpenalized} LMM
}
\keyword{internal}
