% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss-plmm.R
\name{loss.plmm}
\alias{loss.plmm}
\title{Loss method for "plmm" class}
\usage{
loss.plmm(y, yhat)
}
\arguments{
\item{y}{Observed response vector}

\item{yhat}{Predicted response vector}
}
\description{
Loss method for "plmm" class
}
\examples{
fit <- plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X), intercept = FALSE)
head(loss.plmm(yhat = (fit$SUy), y = admix$y))
#TODO (Aug. 18, 2022): ensure that the above choice of 'yhat' is sensible 
}
