% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{MCP}
\alias{MCP}
\title{helper function to implement MCP penalty
The helper functions to implement each penalty.}
\usage{
MCP(z, l1, l2, gamma, v)
}
\arguments{
\item{z}{a vector representing the solution over active set at each feature}

\item{l1}{upper bound (on beta)}

\item{l2}{lower bound (on beta)}

\item{gamma}{The tuning parameter of the MCP penalty}

\item{v}{the 'xtx' term}
}
\value{
numeric vector of the MCP-penalized coefficient estimates within the given bounds
}
\description{
helper function to implement MCP penalty
The helper functions to implement each penalty.
}
\keyword{internal}
