% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdr_plot.R
\name{mfdr_plot}
\alias{mfdr_plot}
\title{Plot method for mFDR}
\usage{
mfdr_plot(mfdr, logscale = TRUE, ...)
}
\arguments{
\item{mfdr}{A data frame as returned by \code{mfdr()}}

\item{logscale}{Logical: should lambda values be plotted on the log scale? Defaults to TRUE.}

\item{...}{Other arguments passed to \code{plot()}}
}
\description{
Plot method for mFDR
}
\examples{
cv <- cv.plmm(admix$X, admix$y)
fit <- cv$fit
mfdr_plot(mfdr(fit), type = "l")
}
