% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_data.R
\name{mode}
\alias{mode}
\title{Helper function to calculate the mode}
\usage{
mode(x)
}
\arguments{
\item{x}{A numeric vector. May have missing values.}
}
\value{
Numeric value with a mode of \code{x}. Note that if \code{x} is multi-modal,
the value returned with just choose one of those modes!
}
\description{
Helper function to calculate the mode
}
