% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_and_count_bigsnp.R
\name{name_and_count_bigsnp}
\alias{name_and_count_bigsnp}
\title{A helper function to label and summarize the contents of a \code{bigSNP}}
\usage{
name_and_count_bigsnp(obj, id_var, quiet, outfile)
}
\arguments{
\item{obj}{a \code{bigSNP} object, possibly subset by \code{add_external_phenotype()}}

\item{id_var}{String specifying which column of the PLINK \code{.fam} file has the unique sample identifiers. Options are "IID" (default) and "FID".}

\item{quiet}{Logical: should console messages be silenced? Defaults to FALSE}

\item{outfile}{The string with the name of the \code{.log} file}
}
\value{
a list with 7 components:
\itemize{
\item \code{na_counts}: vector of missing SNP counts in \code{genotypes}
\item \code{obj}: a modified \code{bigSNP} list with additional components
\item \code{og_plink_ids}: either the IID or FID column from \code{.fam}, determined by \code{id_var}
\item \code{chr}: p-length containing the chromosomes for each SNP
\item \code{X}: the \code{obj$genotypes} as its own FBM
\item \code{pos}: vector of physical positions of the SNPs
\item \code{chr_range}: vector containing the minimum and maximum values of \code{chr}.
Character strings are treated as the \strong{maximum}.
}
}
\description{
A helper function to label and summarize the contents of a \code{bigSNP}
}
\keyword{internal}
