% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_example.R
\name{plink_example}
\alias{plink_example}
\title{A function to help with accessing example PLINK files}
\usage{
plink_example(path = NULL, parent = FALSE)
}
\arguments{
\item{path}{Optional argument specifying a path (filename) for an external data file in \code{inst/extdata}}

\item{parent}{If \code{path=TRUE} and the user wants the name of the parent directory where that file is located, set \code{parent=TRUE}. Defaults to FALSE.}
}
\value{
If \code{path=NULL}, a character vector of file names is returned. If path is given, then a character string
with the full file path
}
\description{
A function to help with accessing example PLINK files
}
\keyword{internal}
