% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm-package.R
\docType{package}
\name{plmm-package}
\alias{plmm-package}
\title{plmm: Fit nonconvex-penalized linear mixed models to account for the presence of unobserved confounding effects}
\description{
Fit a penalized linear mixed model that corrects for unobserved confounding factors. plmm infers and corrects for the presence of unobserved confounding effects such as population stratification and environmental heterogeneity. It then fits a linear model via lasso penalized maximum likelihood. Originally designed for the multivariate analysis of SNP data, plmm eliminates the need for LD pruning and subpopulation-specific analyses. Functions for the appropriate processing of PLINK files are also supplied.
}
\author{
\strong{Maintainer}: Tabitha K. Peter \email{tabitha-peter@uiowa.edu}

Authors:
\itemize{
  \item Anna C. Reisetter \email{anna-reisetter@uiowa.edu} (\href{https://orcid.org/0000-0001-8332-4585}{ORCID})
  \item Patrick J. Breheny (\href{https://orcid.org/0000-0002-0650-1119}{ORCID})
  \item Yujing Lu
}

}
\keyword{internal}
