% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm.R
\name{plmm}
\alias{plmm}
\title{Fit a linear mixed model via penalized maximum likelihood.}
\usage{
plmm(
  design,
  y = NULL,
  K = NULL,
  eta = NULL,
  penalty = "lasso",
  init = NULL,
  gamma,
  alpha = 1,
  lambda_min,
  nlambda = 100,
  lambda,
  eps = 1e-04,
  max_iter = 10000,
  dfmax = NULL,
  warn = TRUE,
  trace = FALSE,
  save_rds = NULL,
  return_fit = TRUE,
  restandardize = TRUE,
  ...
)
}
\arguments{
\item{design}{The first argument must be one of three things:
(1) \code{plmm_design} object (as created by \code{create_design()})
(2) a string with the file path to a design object (the file path must end in \code{.rds})
(3) a \code{matrix} or \code{data.frame} object representing the design matrix of interest.}

\item{y}{Optional: In the case where \code{design} is a \code{matrix} or \code{data.frame}, the user must also supply
a numeric outcome vector as the \code{y} argument. In this case, \code{design} and \code{y} will be passed
internally to \code{create_design(X = design, y = y)}.}

\item{K}{Similarity matrix used to rotate the data. This should either be:
(1) a known matrix that reflects the covariance of y,
(2) an estimate (Default is \eqn{\frac{1}{p}(XX^T)}), or
(3) a list with components \code{s} and \code{U}, as returned by a previous \code{plmm()} model fit on the same data.
\cr \strong{Note}: If a user provides their own \code{K} matrix, it is decomposed as provided and will \emph{not} be scaled. User-provided K functionality is currently not supported for filebacked data.}

\item{eta}{Optional argument to input a specific eta term rather than estimate it from the data. If K is a known covariance matrix that is full rank, this should be 1.}

\item{penalty}{The penalty to be applied to the model. Either "lasso" (the default), "SCAD", or "MCP".}

\item{init}{Initial values for coefficients. Default is 0 for all columns of X.}

\item{gamma}{The tuning parameter of the MCP/SCAD penalty (see details). Default is 3 for MCP and 3.7 for SCAD.}

\item{alpha}{Tuning parameter for the Mnet estimator which controls the relative contributions from the MCP/SCAD penalty and the ridge, or L2 penalty. \code{alpha = 1} is equivalent to MCP/SCAD penalty, while \code{alpha = 0} would be equivalent to ridge regression.
However, \code{alpha = 0} is not supported; alpha may be arbitrarily small, but not exactly 0.}

\item{lambda_min}{The smallest value for lambda, as a fraction of the maximum lambda. Default is .001 if the number of observations is larger than the number of covariates and .05 otherwise.}

\item{nlambda}{Length of the sequence of lambda. Default is 100.}

\item{lambda}{A user-specified sequence of lambda values. By default, a sequence of values of length \code{nlambda} is computed, equally spaced on the log scale.}

\item{eps}{Convergence threshold. The algorithm iterates until the RMSE for the change in linear predictors for each coefficient is less than \code{eps}. Default is \code{1e-4}.}

\item{max_iter}{Maximum number of iterations (total across entire path). Default is 10000.}

\item{dfmax}{Maximum number of non-zero coefficients that may enter the model. Default is NULL (no maximum)}

\item{warn}{Return warning messages for failures to converge and model saturation? Default is TRUE.}

\item{trace}{If set to TRUE, inform the user of progress by announcing the beginning of each step of the modeling process. Default is FALSE.}

\item{save_rds}{Optional: if a filepath and name \emph{without} the \code{.rds} suffix is specified (e.g., \code{save_rds = "~/dir/my_results"}), then the model results are saved to the provided location (e.g., "~/dir/my_results.rds").
Accompanying the RDS file is a log file for documentation, e.g., "~/dir/my_results.log".
Defaults to NULL, which does not save any RDS or log files.}

\item{return_fit}{Optional: a logical value indicating whether the fitted model should be returned as a \code{plmm} object in the current (assumed interactive) session.
Defaults to TRUE.}

\item{restandardize}{Should the X matrix be restandardized after rotation? Default is TRUE.}

\item{...}{Additional optional arguments to \code{plmm_checks()}}
}
\value{
A list which includes 18 items:
\itemize{
\item \code{beta_vals}: The matrix of estimated coefficients. Rows are predictors (with the first row being the intercept), and columns are values of \code{lambda}.
\item \code{std_Xbeta}: A matrix of the linear predictors on the scale of the standardized design matrix. Rows are predictors, columns are values of \code{lambda}.
\strong{Note}: \code{std_Xbeta} will not include rows for the intercept or for constant features.
\item \code{std_X_details}: A list with 9 items:
\itemize{
\item \code{center}: The center values used to center the columns of the design matrix
\item \code{scale}: The scaling values used to scale the columns of the design matrix
\item \code{ns}: An integer vector of the nonsingular columns of the original data
\item \code{unpen}: An integer vector of indices of the unpenalized features, if any were specified in the design
\item \code{unpen_colnames}: A character vector of the column names of any unpenalized features.
\item \code{X_colnames}: A character vector with the column names of all features in the original design matrix
\item \code{X_rownames}: A character vector with the row names of all features in the original design matrix; if none were provided, these are named 'row1', 'row2', etc.
\item \code{std_X_colnames}: A subset of \code{X_colnames} representing only nonsingular columns (i.e., the columns indexed by \code{ns})
\item \code{std_X_rownames}: A subset of \code{X_rownames} representing rows that passed QC filtering & and are represented in both the genotype and phenotype data sets (this only applies to PLINK data)
}
\item \code{std_X}: If design matrix is filebacked, the descriptor for the filebacked data is returned using \code{bigmemory::describe()}. If the the data were stored in-memory, nothing is returned (\code{std_X} is NULL).
\item \code{y}: The outcome vector used in model fitting.
\item \code{p}: The total number of columns in the design matrix (including any singular columns, excluding the intercept).
\item \code{plink_flag}: Logical - did the data come from PLINK files?
\item \code{lambda}: A numeric vector of the tuning parameter values used in model fitting.
\item \code{eta}: A double between 0 and 1 representing the estimated proportion of the variance in the outcome attributable to population/correlation structure.
\item \code{penalty}: A character string indicating the penalty with which the model was fit (e.g., 'MCP')
\item \code{gamma}: A numeric value indicating the tuning parameter used for the SCAD or MCP penalties. Not relevant for lasso models.
\item \code{alpha}: A numeric value indicating the elastic net tuning parameter.
\item \code{loss}: A vector with the numeric values of the loss at each value of \code{lambda} (calculated on the ~rotated~ scale)
\item \code{penalty_factor}: A vector of indicators corresponding to each predictor, where 1 = predictor was penalized.
\item \code{ns_idx}: An integer vector with the indices of predictors which were non-singular features (i.e., features which had variation), where feature 1 is the intercept.
\item \code{iter}: An integer vector with the number of iterations needed in model fitting for each value of \code{lambda}
\item \code{converged}: A vector of logical values indicating whether the model fitting converged at each value of \code{lambda}
\item \code{K}: a list with 2 elements, \code{s} and \code{U} ---
\itemize{
\item \code{s}: a vector of the non-zero eigenvalues of the relatedness matrix K (note: K is the kinship matrix for genetic/genomic data; see the article on notation for details)
\item \code{U}: a matrix of the eigenvectors of K associated with \code{s}
}
}
}
\description{
Fit a linear mixed model via penalized maximum likelihood.
}
\examples{
# using admix data
fit <- plmm(admix$X, admix$y)
s <- summary(fit, idx = 50)
print(s)
plot(fit)
}
