% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_checks.R
\name{plmm_checks}
\alias{plmm_checks}
\title{A function to perform checks on passed objects before model fitting.}
\usage{
plmm_checks(
  design,
  K = NULL,
  eta = NULL,
  penalty = "lasso",
  init = NULL,
  gamma,
  alpha = 1,
  dfmax = NULL,
  trace = FALSE,
  save_rds = NULL,
  return_fit = TRUE,
  ...
)
}
\arguments{
\item{design}{The design object, as created by \code{create_design()}}

\item{K}{Similarity matrix used to rotate the data. This should either be
(1) a known matrix that reflects the covariance of y,
(2) an estimate (Default is \eqn{\frac{1}{p}(XX^T)}), or
(3) a list with components \code{s} and \code{U}, as returned by a previous \code{plmm()} model fit on the same data.}

\item{eta}{Optional argument to input a specific eta term rather than estimate it from the data. If K is a known covariance matrix that is full rank, this should be 1.}

\item{penalty}{The penalty to be applied to the model. Either "MCP" (the default), "SCAD", or "lasso".}

\item{init}{Initial values for coefficients. Default is 0 for all columns of X.}

\item{gamma}{The tuning parameter of the MCP/SCAD penalty (see details). Default is 3 for MCP and 3.7 for SCAD.}

\item{alpha}{Tuning parameter for the Mnet estimator which controls the relative contributions from the MCP/SCAD penalty and the ridge, or L2 penalty. \code{alpha = 1} is equivalent to MCP/SCAD penalty, while \code{alpha = 0} would be equivalent to ridge regression.
However, \code{alpha = 0} is not supported; alpha may be arbitrarily small, but not exactly 0.}

\item{dfmax}{Maximum number of non-zero coefficients that may enter the model. Default is NULL (no maximum)}

\item{trace}{If set to TRUE, inform the user of progress by announcing the beginning of each step of the modeling process. Default is FALSE.}

\item{save_rds}{Optional: if a filepath and name is specified (e.g., \code{save_rds = "~/dir/my_results.rds"}), then the model results are saved to the provided location. Defaults to NULL, which does not save the result.}

\item{return_fit}{Optional: a logical value indicating whether the fitted model should be returned as a \code{plmm} object in the current (assumed interactive) session. Defaults to TRUE.}

\item{...}{Additional arguments to \code{get_data()}}
}
\value{
A list which includes 16 items:
\itemize{
\item \code{std_X}: The standardized design matrix. If design matrix is filebacked, the descriptor for the filebacked data is returned using \code{bigmemory::describe()}.
\item \code{std_X_details}: Metadata for \code{std_X}.
\item \code{std_X_n}: Number of rows in \code{std_X}.
\item \code{std_X_p}: Number of columns in \code{std_X}.
\item \code{y}: Original outcome vector.
\item \code{y_name}: Variable name of \code{y}.
\item \code{centered_y}: The centered outcome vector.
\item \code{K}: The relationship matrix (as passed by \code{plmm()}, may be NULL)
\item \code{eta}: Estimated proportion of the variance in the outcome attributable to population/correlation structure (as passed by \code{plmm()}, may be NULL)
\item \code{fbm_flag}: Logical, is \code{std_X} filebacked?
\item \code{plink_flag}: Logical, does \code{std_X} originate from PLINK files?
\item \code{penalty}: A character string indicating the penalty type.
\item \code{gamma}: Tuning parameter for the SCAD or MCP penalties.
\item \code{init}: Initialized values for beta coefficients.
\item \code{dfmax}: Maximum number of non-zero coefficients that may enter the model.
\item \code{n}: Number of rows in the original design matrix prior to standardization procedures.
\item \code{p}: Number of columns in the original design matrix prior to standardization procedures.
}
}
\description{
A function to perform checks on passed objects before model fitting.
}
\keyword{internal}
