% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_las.R
\name{plmm_las}
\alias{plmm_las}
\title{Fit a penalizedLMM with lasso regularization using penalizedLMM using various intercept and standardization options}
\usage{
plmm_las(
  X,
  y,
  X_for_K = X,
  p1,
  intercept = TRUE,
  centerY = FALSE,
  centerRtY = FALSE,
  standardizeX = FALSE,
  standardizeRtX = FALSE
)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Continuous outcome vector.}

\item{X_for_K}{X matrix used to compute the similarity matrix, K. For multi-chromosome analysis this may be supplied in order to perform a leave-one-chromosome-out correction. The objective here is to adjust for population stratification and unobserved confounding without rotating out the causal SNP effects.}

\item{p1}{Number of causal SNPs. Lambda will be selected such that <= p1 variables enter the model.}

\item{intercept}{Logical flag for whether an intercept should be included.}

\item{centerY}{Logical flag for whether the outcome variable, y, should be centered and an intercept coefficient estimated based on its mean. Defaults to FALSE.}

\item{centerRtY}{Logical flag for whether the rotated outcome variable, SUy, should be centered and an intercept coefficient estimated based on its mean. Defaults to FALSE.}

\item{standardizeX}{Should standardization be performed on the original X matrix? Defaults to FALSE.}

\item{standardizeRtX}{Should standardization be performed on the rotated X matrix? Defaults to FALSE.}
}
\description{
This function allows you to fit a linear mixed model via penalized maximum likelihood with null model variance component estimation.
}
