% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_nll.R
\name{plmm_nll}
\alias{plmm_nll}
\title{Evaluate the negative log-likelihood of a null Gaussian penalizedLMM model}
\usage{
plmm_nll(eta, Uy, S)
}
\arguments{
\item{eta}{The proportion of variance in the outcome that is attributable to causal SNP effects. In other words, SNR. Sometimes referred to as the narrow-sense heritability.}

\item{Uy}{The the continuous outcome, y, rotated by the eigenvectors of the similarity matrix, K.}

\item{S}{The eigenvalues of the similarity matrix, K.}
}
\description{
This function allows you to evaluate the negtive log-likelihood of a linear mixed model under the assumption of a null model in order to estimate the variance parameter, eta.
}
\examples{
admix$V <- admix$X\%*\%t(admix$X) # create an estimated covariance matrix 
# NB: this is an estimate of K 
ev <- eigen(admix$V)
U <- ev$vectors
fit <- plmm(X = admix$X, y = admix$y, V = admix$V, penalty = "MCP")
(plmm_nll(eta = fit$eta, Uy = U\%*\%admix$y, S = ev$values ))
}
