% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_nll_nonnull.R
\name{plmm_nll_nonnull}
\alias{plmm_nll_nonnull}
\title{Evaluate the negative log-likelihood of a non-null Gaussian plmm model}
\usage{
plmm_nll_nonnull(fit, SUX, SUy, S, eta)
}
\arguments{
\item{fit}{An object of class plmm_fit.}

\item{SUX}{Rotated design matrix including rotated intercept and unpenalized columns, if present.}

\item{SUy}{Rotated outcome vector.}

\item{S}{Eigenvalues from similarity matrix used for model fitting.}

\item{eta}{Estimated $eta$ value from object fit.}
}
\description{
Evaluate the negative log-likelihood of a non-null Gaussian plmm model
}
