% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_prep.R
\name{plmm_prep}
\alias{plmm_prep}
\title{PLMM prep: a function to run checks, eigendecomposition, and rotation prior to fitting a PLMM model}
\usage{
plmm_prep(
  std_X,
  std_X_n,
  std_X_p,
  centered_y,
  penalty_factor,
  K = NULL,
  eta = NULL,
  fbm_flag,
  trace = NULL,
  ...
)
}
\arguments{
\item{std_X}{Column standardized design matrix. May include clinical covariates and other non-SNP data.}

\item{std_X_n}{The number of observations in \code{std_X} (integer)}

\item{std_X_p}{The number of features in \code{std_X} (integer)}

\item{centered_y}{Continuous outcome vector, centered.}

\item{penalty_factor}{A multiplicative factor for the penalty applied to each coefficient.}

\item{K}{Similarity matrix used to rotate the data. This should either be:
(1) a known matrix that reflects the covariance of y,
(2) an estimate (Default is \eqn{\frac{1}{p}(XX^T)}), or
(3) a list with components \code{s} and \code{U}, as returned by a previous \code{plmm()} model fit on the same data.}

\item{eta}{Optional argument to input a specific eta term rather than estimate it from the data. If K is a known covariance matrix that is full rank, this should be 1.}

\item{fbm_flag}{Logical: is \code{std_X} a filebacked \code{big.matrix} object? This is set internally by \code{plmm()}.}

\item{trace}{If set to TRUE, inform the user of progress by announcing the beginning of each step of the modeling process. Default is FALSE.}

\item{...}{Not used}
}
\value{
List with these components:
\itemize{
\item \code{std_X}: Standardized design matrix. If design matrix is filebacked, the descriptor for the filebacked data is returned using \code{bigmemory::describe()}.
\item \code{centered_y}: Vector of centered outcomes
\item \code{K}: Similarity matrix
\item \code{s}: Vector of the non-zero eigenvalues of \code{K}
\item \code{U}: Matrix of eigenvectors of \code{K} associated with \code{s} (same as left singular values of X).
\item \code{eta}: The numeric value of the estimated eta parameter
\item \code{penalty_factor} A multiplicative factor for the penalty applied to each coefficient.
\item \code{incpt_flag} Logical: Does the model require fitting an intercept?
\item \code{trace}: If set to TRUE, inform the user of progress by announcing the beginning of each step of the modeling process
}
}
\description{
This is an internal function for \code{plmm()}
}
\keyword{internal}
