% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_scad.R
\name{plmm_scad}
\alias{plmm_scad}
\title{Fit a penalizedLMM with SCAD regularization using penalizedLMM}
\usage{
plmm_scad(X, y, p1, standardize = FALSE, K)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Continuous outcome vector.}

\item{p1}{Number of causal SNPs. Lambda will be selected such that <= p1 variables enter the model.}

\item{standardize}{Should standardization be performed within \code{glmnet()}? Defaults to FALSE.}

\item{K}{Similarity matrix. For multi-chromosome analysis this may be supplied in order to perform a leave-one-chromosome-out correction. The objective here is to adjust for population stratification and unobserved confounding without rotating out the causal SNP effects.}
}
\description{
This function allows you to fit a linear mixed model via penalized maximum likelihood with null model variance component estimation.
}
\examples{
RRM <- genRelatednessMat(X = scale(admix$X))
fit <- plmm_scad(X = admix$X, y = admix$y, p1 = 5, K = RRM) #FIXME: throws Error in seq.default(log(lambda.max), log(lambda.min * lambda.max), len = nlambda) : 'from' must be a finite number
}
