% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-cv-plmm.R
\name{predict.cv_plmm}
\alias{predict.cv_plmm}
\title{Predict method for \code{cv_plmm} class}
\usage{
\method{predict}{cv_plmm}(
  object,
  newX,
  type = c("blup", "coefficients", "vars", "nvars", "lp"),
  X,
  lambda,
  idx = object$min,
  ...
)
}
\arguments{
\item{object}{An object of class \code{cv_plmm}.}

\item{newX}{Matrix of values at which predictions are to be made (not used for \code{type} = "coefficients", "vars", or "nvars").
This can be either a filebacked \code{big.matrix} or a matrix object. Note: Columns of this argument must be named!}

\item{type}{A character argument indicating what type of prediction should be
returned. Options are "lp," "coefficients," "vars," "nvars," and "blup." See details.}

\item{X}{Optional: if \code{type = 'blup'} and the model was fit in-memory, the design matrix used to fit the model represented in \code{object} must be supplied.
When supplied, this design matrix will be standardized using the center/scale values in \code{object$std_X_details}, so please \strong{do not} standardize this matrix before supplying here.
\strong{Note}: If the model was fit file-backed, then the filepath to the \code{.bk} file with this standardized design matrix is returned as \code{std_X} in the fit supplied to \code{object}.}

\item{lambda}{A numeric vector of regularization parameter \code{lambda} values
at which predictions are requested.}

\item{idx}{Vector of indices of regularization parameter \code{lambda} at which
predictions are requested.  By default, this is the lambda index which minimizes the cross-validation error.}

\item{...}{Additional optional arguments}
}
\value{
Depends on the \code{type} - see Details
}
\description{
Predict method for \code{cv_plmm} class
}
\details{
Define beta-hat as the coefficients estimated at the value of lambda that minimizes cross-validation error (CVE). Then options for \code{type} are as follows:
\itemize{
\item \code{lp} (linear predictor): uses the product of \code{newX} and the beta coefficients of \code{object} to predict new values of the outcome. This does not incorporate the correlation structure of the data.
\item \code{blup} (acronym for Best Linear Unbiased Predictor): adds to the \code{lp} a value that represents the estimated random effect. This addition is a way of incorporating
the estimated correlation structure of data into our prediction of the outcome.
\item \code{coefficients}: returns the estimated beta-hat
\item \code{vars}: returns the \emph{indices} of variables (e.g., SNPs) with nonzero coefficients at each value of lambda. EXCLUDES intercept.
\item \code{nvars}: returns the \emph{number} of variables (e.g., SNPs) with nonzero coefficients at each value of lambda. EXCLUDES intercept.
}
}
\examples{
set.seed(123)
train_idx <- sample(1:nrow(admix$X), 100)
# Note: ^ shuffling is important here! Keeps test and train groups comparable.
train <- list(X = admix$X[train_idx,], y = admix$y[train_idx])
train_design <- create_design(X = train$X, y = train$y)

test <- list(X = admix$X[-train_idx,], y = admix$y[-train_idx])
fit <- cv_plmm(design = train_design)

pred1 <- predict(object = fit, newX = test$X, X = train$X) # Minimum CVE lambda
pred2 <- predict(object = fit, newX = test$X, X = train$X, idx = fit$min1se) # 1 SE lambda
}
