% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-summary-cv-plmm.R
\name{print.summary.cv.plmm}
\alias{print.summary.cv.plmm}
\title{Print method for summary.cv.plmm objects}
\usage{
\method{print}{summary.cv.plmm}(obj, digits, ...)
}
\arguments{
\item{obj}{An object of class \code{summary.cv.plmm}}

\item{digits}{The number of digits to use in formatting output}

\item{...}{}
}
\description{
Print method for summary.cv.plmm objects
}
\examples{
cv_fit <- cv.plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X))
s <- summary.cv.plmm(cv_fit)
print.summary.cv.plmm(s)
}
