% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_matrix.R
\name{process_matrix}
\alias{process_matrix}
\title{A function to process data from an in-memory matrix}
\usage{
process_matrix(X, rds_dir, rds_prefix, quiet = FALSE)
}
\arguments{
\item{X}{A numeric matrix of data. Note: rows should be observations (e.g., samples); columns should be features (e.g., predictors, SNPs, variants, etc.)}

\item{rds_dir}{String specifying a filepath for a directory where output will be created}

\item{rds_prefix}{String specifying the prefix of the '.rds' file to be created}

\item{quiet}{Logical; if true, a messsage will be printed indicating the location of the processed data.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      The complete filepath (including '.rds' extension) of the newly created object
}\if{html}{\out{</div>}}
}
\description{
A function to process data from an in-memory matrix
}
\examples{
process_matrix(X = admix$X, rds_dir = tempdir(), rds_prefix = 'admix')
}
