% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_plink_2.R
\name{process_plink_2}
\alias{process_plink_2}
\title{An updated function for processing PLINK files}
\usage{
process_plink_2(
  data_dir,
  prefix,
  rds = FALSE,
  impute = "simple",
  quiet = FALSE,
  gz = FALSE,
  ...
)
}
\arguments{
\item{data_dir}{The path to the bed/bim/fam data files}

\item{prefix}{The prefix (as a character string) of the bed/fam data files}

\item{rds}{Logical: does an rds file for these data files already exist in \code{data_dir}? Defaults to FALSE}

\item{quiet}{Logical: should messages be printed to the console? Defaults to TRUE}

\item{gz}{Logical: are the bed/bim/fam files g-zipped? Defaults to FALSE. NOTE: if TRUE, process_plink will unzip your zipped files.}

\item{...}{Other arguments to bigsnpr::snp_fastImpute or bigsnpr::snp_fastImputeSimple (depending on choice of \code{impute})
#' Note: an 'impute' argument is still under construction. Only "simple" method is available at this time, but we hope to add "xgboost" as an option in the future.}
}
\value{
A list of two components:
\itemize{
\item X: the fully-imputed design matrix
\item constants_idx: A numeric vector with the indices of the constant features. An index of 5 indicates that the 5th feature is constant (i.e. has zero variance)
}
}
\description{
An updated function for processing PLINK files
}
\examples{
\dontrun{
cad_lite <- process_plink_2(data_dir = plink_example(path = "cad_lite.bed", parent = T), prefix = "cad_lite", rds = F, gz = TRUE)
}
}
\keyword{internal}
