% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_plink.R
\name{process_plink_snpStats}
\alias{process_plink_snpStats}
\title{This function allows you to preprocess PLINK bed/bim/fam files using \code{snpStats}, an older package compared to \code{bigsnpr}. Unreliable SNPs are removed and missing values are imptued using either \code{snpStats}, or if not tagged, the HWE mean value.}
\usage{
process_plink_snpStats(
  prefix,
  dataDir,
  gz = FALSE,
  sexcheck = FALSE,
  na.strings = "-9",
  impute = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{prefix}{Character argument that is the prefix of your bed/bim/fam files.}

\item{dataDir}{Directory where plink files (and .sexcheck files are located if \code{sexcheck = TRUE}) are located}

\item{gz}{A logical: are the PLINK files g-zipped (i.e. are the file endings '.gz')?. Defaults to FALSE. NOTE: if TRUE, process_plink_snpStats will unzip your zipped files}

\item{sexcheck}{Logical flag for whether or not PLINK sexcheck files should be incorporated. Defaults to FALSE. If TRUE, sexcheck files must be of the form "prefix.sexcheck" in the same directory as the bed/fam files}

\item{na.strings}{For \code{snpStats}. Strings in .bam and .fam files to be recoded as NA. Defaults to "-9"}

\item{impute}{Logical flag for whether imputation should be performed. Defaults to TRUE since plmm cannot handle missing values.}

\item{quiet}{Logical flag: should progress messages be printed? Defaults to FALSE.}
}
\value{
A three element list object:
\itemize{
\item \code{genotypes} The filtered and imputed genotypes in a snpMatrix object with subjects in rows and SNPs in columns. This is a numeric matrix.
\item \code{map} A matrix of SNP data.
\item \code{fam} A matrix of subject data.
}
}
\description{
This function allows you to preprocess PLINK bed/bim/fam files using \code{snpStats}, an older package compared to \code{bigsnpr}. Unreliable SNPs are removed and missing values are imptued using either \code{snpStats}, or if not tagged, the HWE mean value.
}
\examples{
\dontrun{
# the output of calls to 'plink_example' will vary according to the users' directory structure
test <- process_plink_snpStats(prefix = "cad_lite", dataDir = plink_example(path="cad_lite.fam.gz", parent=T), gz = TRUE)
test2 <- process_plink_snpStats(prefix = "cad_mid", dataDir = plink_example(path="cad_mid.fam.gz", parent=T))
}


}
\keyword{internal}
