% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_X.R
\name{process_X}
\alias{process_X}
\title{A function to read in large data files as an FBM}
\usage{
process_X(
  file,
  data_dir,
  rds_dir = data_dir,
  ind.col,
  impute = TRUE,
  impute_method = "mean",
  outfile,
  overwrite = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{file}{}

\item{rds_dir}{}

\item{ind.col}{}

\item{outfile}{}

\item{overwrite}{}

\item{quiet}{}

\item{...}{}
}
\description{
A function to read in large data files as an FBM
}
\examples{

\dontrun{
process_X(file = "colon2.txt",
 data_dir = get_example_data(parent = T),
 rds_dir = "~/tmp_files",
  ind.col = 2:2002)


}


}
