% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.plmm}
\alias{residuals.plmm}
\title{Extract residuals from a PLMM fit}
\usage{
\method{residuals}{plmm}(
  object,
  lambda,
  which = 1:length(object$lambda),
  drop = TRUE,
  unrotate = FALSE,
  ...
)
}
\arguments{
\item{object}{Object of class \code{plmm}.}

\item{lambda}{Optional numeric value(s) of the regularization parameter at which residuals are requested (numeric vector). For values of lambda not in the sequence of fitted models, linear interpolation is used.}

\item{which}{Index of the penalty parameter at which residuals are requested (default = all indices). If \code{lambda} is specified, this take precedence over \code{which}.}

\item{drop}{By default, if a single value of lambda is supplied, a vector of residuals is returned (logical; default=\code{TRUE}). Set \code{drop=FALSE} if you wish to have the function always return a matrix (see \code{\link[=drop]{drop()}}).}

\item{unrotate}{Logical: should residuals be 'unrotated', i.e. transformed back to the original scale?}

\item{...}{Not used.}
}
\value{
a numeric matrix with the residuals on the \strong{transformed scale}, where
rows correspond to observations (e.g., samples) and columns correspond to
values of \code{lambda} at which the model was fit.
}
\description{
Currently, only deviance residuals are supported.
}
\examples{
\donttest{
fit <- plmm(admix$X, admix$y)
residuals.plmm(fit)[1:5, 1:5]
head(residuals.plmm(fit, which = 50))
}

}
