% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{a function to rotate data (or 'precondition' data) onto the transformed scale}
\usage{
rotate(s, U, eta, y = NULL, X = NULL)
}
\description{
a function to rotate data (or 'precondition' data) onto the transformed scale
}
