% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_filebacked.R
\name{rotate_filebacked}
\alias{rotate_filebacked}
\title{A function to rotate filebacked data}
\usage{
rotate_filebacked(prep, y, tocenter = TRUE, ...)
}
\arguments{
\item{prep}{The object returned by \code{plmm_prep()}}

\item{y}{The continuous outcome vector.}

\item{tocenter}{Should the matrix be centered in addition to scaled? Defaults to TRUE}

\item{...}{Not used}
}
\value{
a list with 4 items:
\itemize{
\item \code{stdrot_X}: \code{X} on the rotated and re-standardized scale
\item \code{rot_y}: \code{y} on the rotated scale (a numeric vector)
\item \code{stdrot_X_center}: numeric vector of values used to center \code{rot_X}
\item \code{stdrot_X_scale}: numeric vector of values used to scale \code{rot_X}
}
}
\description{
A function to rotate filebacked data
}
\keyword{internal}
