% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_environ_eff.R
\name{sim_environ_eff}
\alias{sim_environ_eff}
\title{Simulate a vector of environmental confounding effects}
\usage{
sim_environ_eff(structureGamma, J)
}
\arguments{
\item{structureGamma}{A character argument describing the desired structure.}

\item{J}{Number of subpopulations.}
}
\description{
This function allows you to simulate an environmental confounding effect. This is used by \code{sim_ps_dat()}, which scales the generated vector in order to comply with the user specified desired variance in the outcome attributable to the environmental effect.
}
\examples{
sim_environ_eff(structureGamma = "linear_concordant", J = 3)
}
