% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_data.R
\name{simple_imputation}
\alias{simple_imputation}
\title{A helper function to impute the mean, median, or mode value}
\usage{
simple_imputation(x, impute_method)
}
\arguments{
\item{x}{A vector of numeric values, typically with some NAs}

\item{impute_method}{Three options here:
\itemize{
\item 'mean' = fill in missing values with the mean
\item 'median' = fill in missing values with the median
\item 'mode' = fill in missing values with a mode
Note: 'the' mode may not exist; x could have multiple modes
}}
}
\value{
a numeric vector with no missing values
}
\description{
A helper function to impute the mean, median, or mode value
}
\examples{
x <- c(sample(1:3, 10, replace = T), NA_real_)
simple_imputation(x, "mean")

}
\keyword{internal}
