% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_data.R
\name{simple_imputation_matrix}
\alias{simple_imputation_matrix}
\title{A helper function to appply \code{simple_imputation} to a matrix}
\usage{
simple_imputation_matrix(X, impute_method)
}
\arguments{
\item{X}{A numeric matrix}

\item{impute_method}{Three options here (to pass to \code{simple_imputation()}):
\itemize{
\item 'mean' = fill in missing values with the mean
\item 'median' = fill in missing values with the median
\item 'mode' = fill in missing values with a mode
Note: 'the' mode may not exist; x could have multiple modes
}}
}
\value{
numeric matrix with imputed values
}
\description{
A helper function to appply \code{simple_imputation} to a matrix
}
\keyword{internal}
