% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_fbm.R
\name{standardize_fbm}
\alias{standardize_fbm}
\title{A helper function to standardize an \code{FBM}}
\usage{
standardize_fbm(subset_X, prefix, rds_dir, ns, non_gen, outfile, quiet)
}
\arguments{
\item{subset_X}{An FBM}

\item{prefix}{The data file name without the extension}

\item{rds_dir}{The directory where the user wants to create the '.rds' and '.bk' files}

\item{ns}{Integer vector with indices of non-singular columns of \code{subset_X}}

\item{outfile}{Optional: the name (character string) of the prefix of the logfile to be written.}

\item{quiet}{Logical: should the messages printed to the console be silenced? Defaults to FALSE.}

\item{unpen}{Integer vector of indices with non-genomic covariates}
}
\value{
A list that includes:
\itemize{
\item 'std_X' - the standardized FBM
\item 'center' - a vector with  the values used to center the columns
\item 'scale' - a vector with the values used to scale the columns
}
}
\description{
A helper function to standardize an \code{FBM}
}
\keyword{internal}
