% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_filebacked.R
\name{standardize_filebacked}
\alias{standardize_filebacked}
\title{A helper function to standardize a filebacked matrix}
\usage{
standardize_filebacked(X, outfile, quiet, tocenter = TRUE)
}
\arguments{
\item{X}{A \code{big.matrix} object that has been subset &/or had any additional predictors appended as columns}

\item{outfile}{Optional: the name (character string) of the logfile to be written.}

\item{quiet}{Logical: should console messages be silenced? Defaults to FALSE}

\item{tocenter}{Should the matrix be centered in addition to scaled? Defaults to TRUE.}
}
\value{
A list with a component called \code{std_X} - this is an FBM with column-standardized data.
List also includes several other indices/meta-data on the standardized matrix
}
\description{
A helper function to standardize a filebacked matrix
}
\keyword{internal}
