% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_in_memory.R
\name{standardize_in_memory}
\alias{standardize_in_memory}
\title{A helper function to standardize matrices}
\usage{
standardize_in_memory(X, tocenter = TRUE)
}
\arguments{
\item{X}{A matrix}

\item{tocenter}{Should the matrix be centered in addition to scaled? Defaults to TRUE.}
}
\value{
A list containing the standardized \code{X} matrix and associated metadata
}
\description{
A helper function to standardize matrices
}
\details{
This function is adapted from https://github.com/pbreheny/ncvreg/blob/master/R/std.R
NOTE: this function returns a matrix \strong{in memory}. For standardizing filebacked
data, use \code{standardize_filebacked()} -- see \code{src/big_standardize.cpp}
}
\keyword{internal}
