% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-cv-plmm.R
\name{summary.cv.plmm}
\alias{summary.cv.plmm}
\title{A summary function for cv.plmm objects}
\usage{
\method{summary}{cv.plmm}(obj, lambda = "min")
}
\arguments{
\item{obj}{A \code{cv.plmm} object}

\item{lambda}{The regularization parameter value at which inference should be reported. Can choose a numeric value, 'min', or '1se'. Defaults to 'min.'}
}
\value{
The return value is an object with S3 class \code{summary.cv.plmm}. The class has its own print method and contains the following list elements:
\itemize{
\item \code{lambda.min}: The lambda value at the minimum cross validation error
\item \code{lambda.1se}: The maximum lambda value within 1 standard error of the minimum cross validation error
\item \code{nvars}: The number of non-zero coefficients at the selected lambda value
\item \code{cve}: The cross validation error at all folds
\item \code{min}: The minimum cross validation error
\item \code{fit}: The \code{plmm} fit used in the cross validation
\item \code{bias}: The mean bias of the cross validation
\item \code{loss}: The loss (at each fold?) TODO: double-check this
\item \code{penalty}: The penalty applied to the fitted model
}
}
\description{
A summary function for cv.plmm objects
}
\examples{
cv_fit <- cv.plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X))
s <- summary.cv.plmm(cv_fit)
print.summary.cv.plmm(s)
}
