% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eta.R
\name{test_eta_estimation}
\alias{test_eta_estimation}
\title{a function to write tests of estimate_eta}
\usage{
test_eta_estimation(
  sig_s,
  sig_eps,
  K,
  intercept = TRUE,
  center_y = FALSE,
  prop_skew = 0,
  return_y = FALSE,
  ...
)
}
\arguments{
\item{sig_s}{Variance attributable to structure}

\item{sig_eps}{Variance of random error}

\item{K}{Matrix to use as relatedness matrix.}

\item{intercept}{Logical: should the outcome be simulated with an intercept? Defaults to TRUE.}

\item{center_y}{Logical: should y be centered? Defaults to FALSE. (Note: this option only makes sense if intercept = TRUE)}

\item{prop_skew}{Proportion (value between 0 and 1) that indicates the probability of an outcome value being skewed. Skewed outcomes are generate from a Chi-squared distribution (i.e., the square of the model's normal distribution).}

\item{return_y}{Logical: should simulated outcome values be returned? Defaults to FALSE.}

\item{...}{Additional args to pass into \code{estimate_eta()}}
}
\description{
a function to write tests of estimate_eta
}
\keyword{internal}
