% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_beta}
\alias{update_beta}
\title{a function to update beta}
\usage{
update_beta(ind, lam, multiplier, alpha, b, z, gamma, v, penalty)
}
\arguments{
\item{ind}{Indices of features (e.,g., active features)}

\item{lam}{A single value of lambda}

\item{alpha}{Tuning paramter controlling the ridge component of penalty, as
in \code{ncvreg()}; default is 1 (meaning no ridge penalty)}

\item{b}{Beta coefficient values}

\item{z}{The solution over the active set}

\item{gamma}{Tuning parameter of the MCP/SCAD penalty, as in \code{ncvreg()};
default is 3 for MCP and 3.7 for SCAD.}

\item{v}{}
}
\description{
a function to update beta
}
