% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_r}
\alias{update_r}
\title{a function to update r}
\usage{
update_r(r, X, shift, ind)
}
\arguments{
\item{r}{Vector of residuals}

\item{X}{A file-backed data frame}

\item{shift}{A vector of the same length of \code{r}, indicating how much to shift/change \code{X}}

\item{ind}{Indicies for which features to use in update (e.g., indicies for active set)}
}
\description{
a function to update r
}
