% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_hat.R
\name{v_hat}
\alias{v_hat}
\title{a function to create the estimated variance matrix from a PLMM fit}
\usage{
v_hat(fit, K = NULL)
}
\arguments{
\item{fit}{An object returned by \code{plmm()}}

\item{K}{An optional matrix as list returned by \code{choose_K()}}
}
\value{
Vhat, a matrix representing the estimated variance
}
\description{
a function to create the estimated variance matrix from a PLMM fit
}
