% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varp.R
\name{varp}
\alias{varp}
\title{Calculate the population variance}
\usage{
varp(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
This function allows you to calculate the population variance; we assume our sample is the population.
}
\examples{
v <- rnorm(5)
varp(v)
}
