% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admix}
\alias{admix}
\title{Semi-simulated SNP data}
\format{
A list with 3 components
\describe{
\item{X}{SNP matrix (197 observations of 100 SNPs)}
\item{y}{vector of simulated (continuous) outcomes}
\item{race}{vector with racial group categorization: # 0 = African, 1 = African American, 2 = European, 3 = Japanese}
}
}
\source{
\url{https://hastie.su.domains/CASI/}
}
\usage{
admix
}
\description{
A dataset containing the 100 SNPs, a demographic variable representing race,
and a simulated outcome
}
\keyword{datasets}
