% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_famfile_ids.R
\name{align_ids}
\alias{align_ids}
\title{A helper function to support \code{process_X()}}
\usage{
align_ids(rownames_X, id_var, quiet, add_predictor)
}
\arguments{
\item{rownames_X}{A character vector with the rownames or IDs of the observations in the filebacked matrix of data}

\item{id_var}{character vector with the same values as \code{rownames_X}}

\item{quiet}{Logical: should a message be printed?}

\item{add_predictor}{External data to include in design matrix. This is the add_predictors... arg in \code{process_X()}}
}
\value{
An object of the same type as add_predictor
}
\description{
A helper function to support \code{process_X()}
}
\keyword{internal}
