% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_ids.R
\name{align_ids}
\alias{align_ids}
\title{A helper function to support \code{process_plink()}
This function is called by another helper, \code{add_predictors_to_bigsnp()}}
\usage{
align_ids(id_var, quiet, add_predictor, og_ids)
}
\arguments{
\item{id_var}{String specifying the variable name of the ID column}

\item{quiet}{Logical: should a message be printed?}

\item{add_predictor}{External data to include in design matrix. This is the add_predictors... arg in \code{process_plink()}}

\item{og_ids}{Character vector with the PLINK ids (FID or IID) from the \emph{original} data (i.e., the data before any subsetting from handling missing phenotypes)}
}
\value{
An object of the same type as add_predictor
}
\description{
A helper function to support \code{process_plink()}
This function is called by another helper, \code{add_predictors_to_bigsnp()}
}
\keyword{internal}
