% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_cbind.R
\name{big_cbind}
\alias{big_cbind}
\title{a version of cbind() for file-backed matrices}
\usage{
big_cbind(A, B, C, quiet)
}
\arguments{
\item{A}{in-memory data}

\item{B}{file-backed data}

\item{C}{file-backed placeholder for combined data}

\item{quiet}{Logical}
}
\value{
C, filled in with all column values of A and B combined
}
\description{
a version of cbind() for file-backed matrices
}
\keyword{internal}
