% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_std.R
\name{big_std}
\alias{big_std}
\title{big_std
A function to standardize a matrix stored as an FBM object}
\usage{
big_std(X, center = NULL, scale, ns = NULL)
}
\arguments{
\item{X}{An file-backed design matrix}

\item{center}{A vector of centering values}

\item{scale}{A vctor of scaling values}

\item{ns}{A vector with the indicies marking the Non-Singular columns of X. Defaults to NULL (meaning all columns are assumed to be nonsingular).}

\item{...}{Other arguments to \code{bigstatsr::big_apply()}}
}
\value{
A standardized matrix of class FBM
}
\description{
big_std
A function to standardize a matrix stored as an FBM object
}
\keyword{internal}
