% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_blup.R
\name{compute_blup}
\alias{compute_blup}
\title{a function to compute the BLUP}
\usage{
compute_blup(fit, Xb, Sigma_21)
}
\arguments{
\item{fit}{An object returned by \code{plmm()}}

\item{Xb}{Linear predictor}

\item{Sigma_21}{Covariance matrix between the training and the testing data. Extracted from \code{estimated_Sigma} that is generated using all observations}
}
\value{
Sigma_hat, a matrix representing the estimated variance
}
\description{
a function to compute the BLUP
}
\keyword{internal}
