% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design.R
\name{create_design}
\alias{create_design}
\title{a function to create a design for PLMM modeling}
\usage{
create_design(
  data_file = NULL,
  rds_dir = NULL,
  X = NULL,
  outcome_col = NULL,
  ...
)
}
\arguments{
\item{data_file}{For filebacked data (data from \code{process_plink()} or \code{process_delim()}), this is the
filepath to the processed data. Defaults to NULL (does not apply for in-memory data)}

\item{rds_dir}{For filebacked data, this is the filepath to the directory/folder where you want the design to be saved.
\strong{Note}: do not include/append the name you want for the to-be-created file -- the name is the argument \code{new_file},
passed to \code{create_design_filebacked()}.}

\item{X}{For in-memory data, this is the design matrix.}

\item{outcome_col}{For in-memory data, this is the numeric vector representing the outcome.}

\item{...}{Additional arguments to pass to \code{create_design_filebacked()} or \code{create_design_in_memory()}.
See the documentation for those helper functions for details.}
}
\value{
An object of class \code{plmm_design}
}
\description{
a function to create a design for PLMM modeling
}
\details{
This function is a wrapper for the other \code{create_design...()} inner functions; all arguments
included here are passed along to the \code{create_design...()} inner function that
matches the type of the data being supplied.
}
\examples{
## Example 1: delimited data ##
# process delimited data
temp_dir <- tempdir()
colon_dat <- process_delim(data_file = "colon2.txt",
 data_dir = find_example_data(parent = TRUE), overwrite = TRUE,
 rds_dir = temp_dir, rds_prefix = "processed_colon2", sep = "\t", header = TRUE)

# prepare outcome data
colon_outcome <- read.delim(find_example_data(path = "colon2_outcome.txt"))

# create a design
colon_design <- create_design(data_file = colon_dat, rds_dir = temp_dir, new_file = "std_colon2",
add_outcome = colon_outcome, outcome_id = "ID", outcome_col = "y",
overwrite = TRUE, logfile = "test.log")

# look at the results
colon_rds <- readRDS(colon_design)
str(colon_rds)

## Example 2: matrix data in-memory
admix_design <- create_design(X = admix$X, outcome_col = admix$y, unpen = "Snp1")

}
