% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_log.R
\name{create_log}
\alias{create_log}
\title{create_log_file}
\usage{
create_log(outfile, ...)
}
\arguments{
\item{outfile}{String specifying the name of the to-be-created file, \emph{without} extension}

\item{...}{Not used}
}
\value{
Nothing is returned, intead a text file with the suffix .log is created.
}
\description{
create_log_file
}
\keyword{internal}
