% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-blups-plmm.R, R/cv-plmm.R
\name{cv.plmm}
\alias{cv.plmm}
\title{Cross-validation for plmm}
\usage{
cv.plmm(
  X,
  y,
  X_for_K = X,
  cv_rotated = TRUE,
  ...,
  cluster,
  nfolds = 10,
  seed,
  fold,
  returnY = FALSE,
  trace = FALSE
)

cv.plmm(
  X,
  y,
  X_for_K = X,
  cv_rotated = TRUE,
  ...,
  cluster,
  nfolds = 10,
  seed,
  fold,
  returnY = FALSE,
  trace = FALSE
)
}
\arguments{
\item{X}{Design matrix. May include clinical covariates and other non-SNP data. If this is the case, X_for_K should be supplied witha  matrix containing only SNP data for computation of GRM.}

\item{y}{Continuous outcome vector.}

\item{X_for_K}{X matrix used to compute the similarity matrix, K. For multi-chromosome analysis this may be supplied in order to perform a leave-one-chromosome-out correction. The objective here is to adjust for population stratification and unobserved confounding without rotating out the causal SNP effects.}

\item{cv_rotated}{Logical flag for whether cross-validation should be done by fitting and predicting using the rotated data (TRUE), or original (FALSE). Defaults to TRUE.}

\item{...}{Additional arguments to plmm}

\item{cluster}{cv.plmm can be run in parallel across a cluster using the parallel package. The cluster must be set up in advance using the makeCluster function from that package. The cluster must then be passed to cv.plmm.}

\item{nfolds}{The number of cross-validation folds. Default is 10.}

\item{seed}{You may set the seed of the random number generator in order to obtain reproducible results.}

\item{fold}{Which fold each observation belongs to. By default the observations are randomly assigned.}

\item{returnY}{Should cv.plmm return the linear predictors from the cross-validation folds? Default is FALSE; if TRUE, this will return a matrix in which the element for row i, column j is the fitted value for observation i from the fold in which observation i was excluded from the fit, at the jth value of lambda.}

\item{trace}{If set to TRUE, inform the user of progress by announcing the beginning of each CV fold. Default is FALSE.}
}
\description{
Performs k-fold cross validation for lasso-, MCP-, or SCAD-penalized penalized linear mixed models over a grid of values for the regularization parameter lambda.

Performs k-fold cross validation for lasso-, MCP-, or SCAD-penalized penalized linear mixed models over a grid of values for the regularization parameter lambda.
}
