% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvf.R
\name{cvf}
\alias{cvf}
\title{Cross-validation internal function for cv.plmm}
\usage{
cvf(i, XX, y, fold, cv.args, predict.args)
}
\arguments{
\item{i}{Fold number to be excluded from fit.}

\item{XX}{Design matrix. May include clinical covariates and other non-SNP data. If this is the case, X_for_K should be supplied witha  matrix containing only SNP data for computation of GRM.}

\item{y}{Continuous outcome vector.}

\item{fold}{n-length vector of fold-assignments.}

\item{cv.args}{List of additional arguments to be passed to plmm.}

\item{predict.args}{List of arguments required for prediction. Varies according to prediction type (response vs. individual).}
}
\description{
Internal function for cv.plmm which calls plmm on a fold subset of the original data.
}
