% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvf.R
\name{cvf}
\alias{cvf}
\title{Cross-validation internal function for cv.plmm}
\usage{
cvf(i, fold, type, cv.args, ...)
}
\arguments{
\item{i}{Fold number to be excluded from fit.}

\item{fold}{n-length vector of fold-assignments.}

\item{type}{A character argument indicating what should be returned from predict.plmm. If \code{type == 'response'} predictions are based on the linear predictor, \code{$X beta$}. If \code{type == 'individual'} predictions are based on the linear predictor plus the estimated random effect (BLUP).}

\item{cv.args}{List of additional arguments to be passed to plmm.}

\item{...}{Optional arguments to \code{predict.list}}
}
\description{
Internal function for cv.plmm which calls plmm on a fold subset of the original data.
}
\keyword{internal}
