% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_K.R
\name{eigen_K}
\alias{eigen_K}
\title{A function to take the eigendecomposition of K
Note: This is faster than taking SVD of X when p >> n}
\usage{
eigen_K(std_X, fbm_flag)
}
\arguments{
\item{std_X}{The \emph{standardized} design matrix, stored as big.matrix object.}

\item{fbm_flag}{Logical: is std_X an FBM obejct? Passed from \code{plmm()}.}
}
\value{
A list with the eigenvectors and eigenvalues of K
}
\description{
A function to take the eigendecomposition of K
Note: This is faster than taking SVD of X when p >> n
}
\keyword{internal}
