% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eta.R
\name{estimate_eta}
\alias{estimate_eta}
\title{Estimate eta (to be used in rotating the data)}
\usage{
estimate_eta(n, s, U, y, incpt_flag)
}
\arguments{
\item{n}{The number of observations}

\item{s}{The non-zero eigenvalues of K, the realized relationship matrix}

\item{U}{The eigenvectors of K associated with s}

\item{y}{Continuous outcome vector}

\item{incpt_flag}{Logical: Does the model require fitting an intercept?}
}
\value{
a numeric value with the estimated value of eta, the variance parameter
}
\description{
This function is called internally by \code{plmm()}
}
\keyword{internal}
