% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eta.R
\name{estimate_eta}
\alias{estimate_eta}
\title{Estimate eta (to be used in rotating the data)
This function is called internally by \code{plmm()}}
\usage{
estimate_eta(n, s, U, y, eta_star)
}
\arguments{
\item{n}{The number of observations}

\item{s}{The singular values of K, the realized relationship matrix}

\item{U}{The left-singular vectors of the \emph{standardized} design matrix}

\item{y}{Continuous outcome vector.}
}
\description{
Estimate eta (to be used in rotating the data)
This function is called internally by \code{plmm()}
}
\keyword{internal}
