% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDataPS.R
\name{genDataPS}
\alias{genDataPS}
\title{Generate data with population structure}
\usage{
genDataPS(
  n = 197,
  p = 98,
  p1 = floor(p/2),
  nJ = c(47, 50, 50, 50),
  structureX = c("admixture", "indep_subpops", "1d_linear", "1d_circular",
    "independent", "other"),
  Fst = NULL,
  inbr = c("heterogeneous", "homogeneous"),
  structureGamma = c("halfandhalf_decreasing_heterogeneous"),
  eta,
  xi,
  standardizeX = TRUE,
  structureX_other = NULL
)
}
\arguments{
\item{n}{Number of observations/samples to simulate.}

\item{p}{Number of SNPs to simulate.}

\item{p1}{Number of SNPs that are causal. Defaults to floor(p/2).}

\item{nJ}{Number of observations in each subpopulation. The length of nJ corresponds to the number of subpopulations.}

\item{structureX}{Type of structure to simulate.}

\item{Fst}{The desired final FST of the admixed individuals. Ranges from 0 to 1. A high Fst implies greater differentiation among populations. Defaults to 0.1 if structureX = 1d_linear and 0.2 if structureX = indep_subpops. Otherwise defaults to NULL.}

\item{inbr}{Indicates whether the desired inbreeding is homogeneous or heterogeneous. Defaults to heterogeneous.}

\item{structureGamma}{The desired structure of the environmental confounding effect. Defaults to halfandhalf_decreasing_heterogeneous.}

\item{eta}{The desired proportion of variance in the outcome that is attributable to causal SNP effects. In other words, SNR.}

\item{xi}{The desired proportion of the non-signal variance in the outcome that is attributable to unobserved environmental confounding effects.}

\item{standardizeX}{Should the generated X matrix be standardized? Defaults to TRUE.}

\item{structureX_other}{If \code{structureX == "other"}, an matrix or SnpMatrix object with subjects in rows and SNPs in columns to be used to generate pseudophenotypes must be supplied here.}
}
\description{
This function allows you to simulate structured genetic data (SNP) with an unobserved environmental confounding effect.
}
