% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness_mat.R
\name{relatedness_mat}
\alias{relatedness_mat}
\title{Generate a relatedness matrix}
\usage{
relatedness_mat(X, ...)
}
\arguments{
\item{X}{Either a numeric matrix of genotypes (subjects in rows, SNPs in columns), or a character prefix for PLINK bed/bim/fam files.}

\item{...}{Optional arguments in the case that a character prefix is specified.}
}
\description{
This function allows you to generate an n by n genetic relatedness matrix. If a numeric matrix is supplied, the RRM (Hayes, 2009) is used
and is computed XX'/p. If a character argument which describes the location and prefix of PLINK bed/bim/bam files is supplied,
a GRM is computed using the GCTA method of SNPrelate, unless another method is specified as an optional argument.
}
\examples{
RRM <- relatedness_mat(X = scale(admix$X))
}
