% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genXps.R
\name{genXps}
\alias{genXps}
\title{Generate and X matrix with population structure}
\usage{
genXps(
  n,
  nJ,
  p,
  structureX = c("admixture", "indep_subpops", "1d_linear", "independent"),
  Fst = NULL,
  inbr = c("homogeneous", "heterogeneous"),
  standardizeX = TRUE,
  plot = FALSE
)
}
\arguments{
\item{n}{Number of observations/samples to simulate.}

\item{nJ}{Number of observations in each subpopulation. The length of nJ corresponds to the number of subpopulations.}

\item{p}{Number of SNPs to simulate.}

\item{structureX}{Type of structure to simulate.}

\item{Fst}{The desired final FST of the admixed individuals. Ranges from 0 to 1. A high Fst implies greater differentiation among populations. Defaults to 0.1 if structureX = 1d_linear and 0.2 if structureX = indep_subpops. Otherwise defaults to NULL.}

\item{inbr}{Indicates whether the desired inbreeding is homogeneous or heterogeneous. Defaults to heterogeneous.}

\item{standardizeX}{Should the generated X matrix be standardized? Defaults to TRUE.}

\item{plot}{Should a plot of the kinship matrix be generated? Defaults to FALSE.}
}
\description{
This function allows you to simulate structured genetic data (SNP).
}
