% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Read in processed data
This function is intended to be called after \code{process_plink()} has been called once.}
\usage{
get_data(path, row_id = NULL, returnX, trace = TRUE)
}
\arguments{
\item{path}{The file path to the RDS object containing the processed data. Do not add the '.rds' extension to the path.}

\item{row_id}{Character string indicating which IDs to use for the rownames of the genotype matrix. Can choose "fid" or "iid", corresponding to the first or second columns in the PLINK .fam file. Defaults to NULL.}

\item{returnX}{Logical: Should the design matrix be returned as a numeric matrix that will be stored in memory. By default, this will be FALSE if the object sizes exceeds 100 Mb.}

\item{trace}{Logical: Should trace messages be shown? Default is TRUE.}
}
\value{
A list with four components:
\itemize{
\item X, the design matrix as either (1) a numeric matrix or (2) a filebacked matrix (FBM). See \code{bigstatsr::FBM()} and \code{bigsnpr::bigSnp-class} documentation for details.
\item fam, a data frame containing the pedigree information (like a .fam file in PLINK)
\item map, a data frame containing the feature information (like a .bim file in PLINK)
\item constants_idx A vector indicating the which columns of X contain constant features (features with no variance).
}
}
\description{
Read in processed data
This function is intended to be called after \code{process_plink()} has been called once.
}
