% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Read in processed data
This function is intended to be called after \code{process_plink()} has been called once.}
\usage{
get_data(path, returnX, trace = TRUE)
}
\arguments{
\item{path}{The file path to the RDS object containing the processed data. Do not add the '.rds' extension to the path.}

\item{returnX}{Logical: Should the design matrix be returned as a numeric matrix that will be stored in memory. By default, this will be FALSE if the object sizes exceeds 100 Mb.}

\item{trace}{Logical: Should trace messages be shown? Default is TRUE.}
}
\value{
A list with these components:
\itemize{
\item std_X, the column-standardized design matrix as either (1) a numeric matrix or (2) a filebacked matrix (FBM). See \code{bigstatsr::FBM()} and \code{bigsnpr::bigSnp-class} documentation for details.
\item fam, a data frame containing the pedigree information (like a .fam file in PLINK)
\item map, a data frame containing the feature information (like a .bim file in PLINK)
\item ns: A vector indicating the which columns of X contain nonsingular features (i.e., features with variance != 0.
\item center: A vector of values for centering each column in X
\item scale: A vector of values for scaling each column in X
}
}
\description{
Read in processed data
This function is intended to be called after \code{process_plink()} has been called once.
}
\details{
The .rds object should have an 'std_X' element - this is what will be used as the design matrix for analysis. This design matrix should \emph{not} include an intercept column (this will be added later in \code{plmm_fit}()).

In the returned list, the \code{fam} data will be sorted by family and by individual, as in \code{dplyr::arrange(family.ID, sample.ID)}.
The rows of \code{X} will be sorted to align in the same order as in \code{fam}, where rownames of \code{X} will be sample ID.
}
\examples{
\dontrun{
pen <- get_data(path = "../tmp_files/penncath_lite", trace = TRUE)
}

}
