% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gic.R
\name{gic}
\alias{gic}
\title{General information criterion method of selecting lambda for "plmm" class}
\usage{
gic(fit, ic = c("bic", "hdbic", "ebic"), std_X, U, s, eta)
}
\arguments{
\item{fit}{An object of class "plmm" where svd_detail has been set to TRUE (the default)}

\item{ic}{Information criterion that should be used to select lambda. Currently supports BIC or HDBIC. Defaults to BIC.}

\item{std_X}{The standardized design matrix}

\item{U}{Optional: Left singular eigenvectors of X}

\item{s}{Optional: Eigenvalues from similarity matrix used for model fitting. If not returned as part of \code{plmm} because \code{returnX == FALSE}, must be supplied explicitly.}

\item{eta}{Optional: Estimate of variance parameter eta}
}
\description{
General information criterion method of selecting lambda for "plmm" class
}
\examples{
fit <- plmm(X = admix$X, y = admix$y)
gic_res <- gic(fit = fit, ic = "bic", std_X = ncvreg::std(admix$X))
names(gic_res)
range(gic_res$gic, na.rm = TRUE) # NAs will result from monomorphic SNPs
}
