% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gic.R
\name{gic}
\alias{gic}
\title{General information criterion method of selecting lambda for "plmm" class}
\usage{
gic(fit, ic = c("bic", "hdbic"), SUX, SUy, S, eta, intercept)
}
\arguments{
\item{fit}{An object of class "plmm."}

\item{ic}{Information criterion that should be used to select lambda. Currently supports BIC or HDBIC. Defaults to BIC.}

\item{SUX}{Rotated design matrix including rotated intercept and unpenalized columns, if present. If not returned as part of \code{plmm} because \code{returnX == FALSE}, must be supplied explicitly.}

\item{SUy}{Rotated outcome vector. If not returned as part of \code{plmm} because \code{returnX == FALSE}, must be supplied explicitly.}

\item{S}{Eigenvalues from similarity matrix used for model fitting. If not returned as part of \code{plmm} because \code{returnX == FALSE}, must be supplied explicitly.}

\item{eta}{Estimated $eta$ value from object fit.}

\item{intercept}{Logical indicating whether an intercept was used in the model}
}
\description{
General information criterion method of selecting lambda for "plmm" class
}
\examples{
\dontrun{
fit <- plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X), intercept = FALSE)
gic(fit = fit, ic = "bic", SUX = fit$SUX, SUy = fit$SUy, S = fit$S, eta = fit$eta)
gic(fit = fit) #FIXME (Aug. 18, 2022)
}
}
