% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{lasso}
\alias{lasso}
\title{helper function to implement lasso penalty}
\usage{
lasso(z, l1, l2, v)
}
\arguments{
\item{z}{solution over active set at each feature}

\item{l1}{upper bound}

\item{l2}{lower bound}

\item{v}{the 'xtx' term}
}
\value{
numeric vector of the lasso-penalized coefficient estimates within the given bounds
}
\description{
helper function to implement lasso penalty
}
