% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso}
\alias{lasso}
\title{Fit a linear model with lasso regularization}
\usage{
lasso(X, y, p1, ...)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Continuous outcome vector.}

\item{p1}{Number of causal SNPs. Lambda will be selected such that <= p1 variables enter the model.}

\item{...}{Additional optional arguments}
}
\description{
This function allows you to fit a linear model via lasso-penalized maximum likelihood with the same output values as those of lmm_lasso_eta. Primarily used for simulation and comparison with lmm_lasso_eta.
}
