% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik}
\alias{log_lik}
\title{Evaluate the negative log-likelihood of an intercept-only Gaussian plmm model}
\usage{
log_lik(eta, n, s, U, y, incpt_flag)
}
\arguments{
\item{eta}{Estimated proportion of the variance in the outcome attributable to population/correlation structure}

\item{n}{The number of observations}

\item{s}{The non-zero eigenvalues of K, the realized relationship matrix}

\item{U}{The eigenvectors of K associated with s}

\item{y}{Continuous outcome vector}

\item{incpt_flag}{Logical: Does the model require fitting an intercept? Passed from \code{estimate_eta}.}
}
\value{
the value of the log-likelihood of the PLMM, evaluated with the supplied parameters
}
\description{
This function allows you to evaluate the negative log-likelihood of a linear mixed model under the assumption of a null model in order to estimate the variance parameter, eta.
}
\keyword{internal}
