% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik}
\alias{log_lik}
\title{Evaluate the negative log-likelihood of an intercept-only Gaussian plmm model}
\usage{
log_lik(eta, n, s, U, y)
}
\arguments{
\item{eta}{The proportion of variance in the outcome that is attributable to causal SNP effects. In other words, SNR. Sometimes referred to as the narrow-sense heritability.}

\item{n}{The number of observations}

\item{s}{The singular values of K, the realized relationship matrix}

\item{U}{The left-singular vectors of the \emph{standardized} design matrix}

\item{y}{Continuous outcome vector.}
}
\description{
This function allows you to evaluate the negative log-likelihood of a linear mixed model under the assumption of a null model in order to estimate the variance parameter, eta.
}
\keyword{internal}
