% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_nonnull.R
\name{logLik_nonnull}
\alias{logLik_nonnull}
\title{Evaluate the negative log-likelihood of a non-null Gaussian plmm model}
\usage{
logLik_nonnull(fit)
}
\arguments{
\item{fit}{An object of class plmm_fit.}

\item{SUX}{Rotated design matrix including rotated intercept and unpenalized columns, if present.}

\item{SUy}{Rotated outcome vector. See pp. 16-18 in A.R.'s thesis for details}

\item{S}{Eigenvalues from similarity matrix used for model fitting.}

\item{eta}{Estimated $eta$ value from object fit.}
}
\description{
Evaluate the negative log-likelihood of a non-null Gaussian plmm model
}
\keyword{internal}
