% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss-plmm.R
\name{loss.plmm}
\alias{loss.plmm}
\title{Loss method for "plmm" class}
\usage{
loss.plmm(y, yhat)
}
\arguments{
\item{y}{Observed response vector}

\item{yhat}{Predicted response vector}
}
\description{
Loss method for "plmm" class
}
\examples{
fit <- plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X))
yhat <- predict(object = fit, newX = admix$X, type = 'lp', lambda = 0.05)
head(loss.plmm(yhat = yhat, y = admix$y))
}
