% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdr.R
\name{mfdr}
\alias{mfdr}
\title{mfdr: Marginal false discovery rates for PLMMs
Based on ncvreg::mfdr()}
\usage{
mfdr(fit)
}
\arguments{
\item{fit}{A \code{plmm} object.}
}
\description{
mfdr: Marginal false discovery rates for PLMMs
Based on ncvreg::mfdr()
}
