% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm.R
\name{plmm}
\alias{plmm}
\title{Fit a linear mixed model with non-convex regularization}
\usage{
plmm(
  X,
  y,
  X_for_K = X,
  penalty = c("MCP", "SCAD", "lasso"),
  penalty.factor = rep(1, ncol(X)),
  gamma,
  alpha = 1,
  lambda.min = ifelse(n > p, 0.001, 0.05),
  nlambda = 100,
  lambda,
  init = rep(0, ncol(X)),
  eps = 1e-04,
  max.iter = 1000,
  convex = TRUE,
  dfmax = p + 1,
  warn = TRUE,
  standardize = TRUE,
  rotation = FALSE,
  returnX = TRUE,
  ...
)
}
\arguments{
\item{X}{Design matrix. May include clinical covariates and other non-SNP data. If this is the case, X_for_K should be supplied witha  matrix containing only SNP data for computation of GRM.}

\item{y}{Continuous outcome vector.}

\item{X_for_K}{X matrix used to compute the similarity matrix, K. For multi-chromosome analysis this may be supplied in order to perform a leave-one-chromosome-out correction. The objective here is to adjust for population stratification and unobserved confounding without rotating out the causal SNP effects.}

\item{rotation}{Logical to indicate whether the weighted rotation of the data should be performed (TRUE), or not (FALSE).
This is primarily for testing purposes and defaults to TRUE.
If \code{FALSE} results should correspond to those from \code{ncvreg} or \code{glmnet}.}
}
\description{
This function allows you to fit a linear mixed model via non-convex penalized maximum likelihood.
}
