% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_checks.R
\name{plmm_checks}
\alias{plmm_checks}
\title{plmm_checks}
\usage{
plmm_checks(
  X,
  fbm = NULL,
  std_needed = NULL,
  col_names = NULL,
  y = NULL,
  k = NULL,
  K = NULL,
  diag_K = NULL,
  eta_star = NULL,
  penalty = c("MCP", "SCAD", "lasso"),
  penalty.factor = NULL,
  init = NULL,
  gamma,
  alpha = 1,
  dfmax = NULL,
  trace = FALSE
)
}
\arguments{
\item{X}{}

\item{fbm}{}

\item{std_needed}{}

\item{col_names}{}

\item{y}{}

\item{k}{}

\item{K}{}

\item{diag_K}{}

\item{eta_star}{}

\item{penalty}{}

\item{penalty.factor}{}

\item{init}{}

\item{gamma}{}

\item{alpha}{}

\item{dfmax}{}

\item{trace}{}
}
\description{
plmm_checks
}
\keyword{internal}
