% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_format.R
\name{plmm_format}
\alias{plmm_format}
\title{PLMM format: a function to format the output of a model constructed with \code{plmm_fit}}
\usage{
plmm_format(fit, X)
}
\arguments{
\item{fit}{A list of parameters describing the output of a model constructed with \code{plmm_fit}}

\item{X}{Design matrix. May include clinical covariates and other non-SNP data.}
}
\value{
A list with the components:
\itemize{
\item beta_vals: The estimated beta values at each value of lambda
\item eta: The estimated eta value
\item lambda: The sequence of lambda values used in model fitting
\item penalty: A string indicating the type of penalty used to fit the model
\item ns_idx: COME BACK HERE
\item iter: The number of iterations at each value of lambda (MAYBE take this out)
\item converged: The convergence status at each value of lambda
}
}
\description{
PLMM format: a function to format the output of a model constructed with \code{plmm_fit}
}
\keyword{internal}
