% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_format.R
\name{plmm_format}
\alias{plmm_format}
\title{PLMM format: a function to format the output of a model constructed with \code{plmm_fit}}
\usage{
plmm_format(
  fit,
  p,
  std_X_details,
  fbm_flag,
  use_feature_names = TRUE,
  feature_names = NULL,
  non_genomic = NULL
)
}
\arguments{
\item{fit}{A list of parameters describing the output of a model constructed with \code{plmm_fit}}

\item{p}{The number of features in the original data (including constant features)}

\item{std_X_details}{A list with 3 items:
\itemize{
\item 'center': the centering values for the columns of \code{X}
\item 'scale': the scaling values for the non-singular columns of \code{X}
\item 'ns': indicies of nonsingular columns in \code{std_X}
}}

\item{fbm_flag}{Logical: is the corresponding design matrix filebacked? Passed from \code{plmm()}.}

\item{use_feature_names}{Logical: should features be named (e.g., should the returned matrix of coefficients have rows/columns named?). Defaults to TRUE.}

\item{feature_names}{A vector of names for features, passed internally if such names are included with the data \code{X} passed to \code{plmm()}}

\item{non_genomic}{Optional vector specifying which columns of the design matrix represent features that are \emph{not} genomic, as these features are excluded from the empirical estimation of genomic relatedness.
For cases where X is a filepath to an object created by \code{process_plink()}, this is handled automatically via the arguments to \code{process_plink()}.
For all other cases, 'non_genomic' defaults to NULL (meaning \code{plmm()} will assume that all columns of \code{X} represent genomic features).}
}
\value{
A list with the components:
\itemize{
\item \code{beta_vals}: the matrix of estimated coefficients on the original scale. Rows are predictors, columns are values of \code{lambda}
\item \code{rotated_scale_beta_vals}: the matrix of estimated coefficients on the ~rotated~ scale. This is the scale on which the model was fit.
\item \code{lambda}: a numeric vector of the lasso tuning parameter values used in model fitting.
\item \code{eta}: a number (double) between 0 and 1 representing the estimated proportion of the variance in the outcome attributable to population/correlation structure.
\item \code{s}: a vectof of the eigenvalues of relatedness matrix \code{K}; see \code{relatedness_mat()} for details.
\item \code{U}: a matrix of the eigenvalues of relatedness matrix \code{K}
\item \code{rot_y}: the vector of outcome values on the rotated scale. This is the scale on which the model was fit.
\item \code{linear_predictors}: the matrix resulting from the product of \code{stdrot_X} and the estimated coefficients on the ~rotated~ scale.
\item \code{penalty}: character string indicating the penalty with which the model was fit (e.g., 'MCP')
\item \code{gamma}: numeric value indicating the tuning parameter used for the SCAD or lasso penalties was used. Not relevant for lasso models.
\item \code{alpha}: numeric value indicating the elastic net tuning parameter.
\item \code{loss}: vector with the numeric values of the loss at each value of \code{lambda} (calculated on the ~rotated~ scale)
\item \code{penalty_factor}: vector of indicators corresponding to each predictor, where 1 = predictor was penalized.
\item \code{ns_idx}: vector with the indicies of predictors which were constant features (i.e., had no variation).
\item \code{p}: the number of features
\item \code{n}: the number of observations (instances)
\item \code{iter}: numeric vector with the number of iterations needed in model fitting for each value of \code{lambda}
\item \code{converged}: vector of logical values indicating whether the model fitting converged at each value of \code{lambda}
}
}
\description{
PLMM format: a function to format the output of a model constructed with \code{plmm_fit}
}
\keyword{internal}
