% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_format.R
\name{plmm_format}
\alias{plmm_format}
\title{PLMM format: a function to format the output of a model constructed with \code{plmm_fit}}
\usage{
plmm_format(fit, convex = TRUE, dfmax = fit$ncol_X + 1, X, K)
}
\arguments{
\item{fit}{A list of parameters describing the output of a model constructed with \code{plmm_fit}}

\item{convex}{convex Calculate index for which objective function ceases to be locally convex? Default is TRUE.}

\item{dfmax}{dfmax Upper bound for the number of nonzero coefficients. Default is no upper bound. However, for large data sets, computational burden may be heavy for models with a large number of nonzero coefficients.}

\item{X}{Design matrix. May include clinical covariates and other non-SNP data.}

\item{K}{Similarity matrix used to rotate the data. This will be passed from plmm() as (1) NULL, in which case relatedness_mat(std(X)) is used, (2) a matrix or (3) a list with components d and u (eigenvalues and eigenvectors, respectively).}
}
\value{
A list with the components:
\itemize{
\item beta_vals: The estimated beta values at each value of lambda
\item eta: The estimated eta value
\item lambda: The sequence of lambda values used in model fitting
\item penalty: A string indicating the type of penalty used to fit the model
\item ns_idx: COME BACK HERE
\item iter: The number of iterations at each value of lambda (MAYBE take this out)
\item converged: The convergence status at each value of lambda
\item X: if returnX = TRUE and size SUX < 100 Mb, the original X will be returned
}
}
\description{
PLMM format: a function to format the output of a model constructed with \code{plmm_fit}
}
\examples{

\dontrun{
# NB: this is an internal function (i.e. it's not exported)
# these examples won't work unless you have penalizedLMM::: 
prep1 <- plmm_prep(X = admix$X, y = admix$y, trace = TRUE)
fit1 <- plmm_fit(prep = prep1)
res1 <- plmm_format(fit1)

}
}
\keyword{internal}
