% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_lasso.R
\name{plmm_lasso}
\alias{plmm_lasso}
\title{Fit a penalizedLMM with lasso regularization using penalizedLMM}
\usage{
plmm_lasso(X, y, p1, ...)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Continuous outcome vector.}

\item{p1}{Number of causal SNPs. Lambda will be selected such that <= p1 variables enter the model.}

\item{...}{Additional optional arguments}
}
\description{
This function allows you to fit a linear mixed model via penalized maximum likelihood with null model variance component estimation.
}
