% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_lasso.R
\name{plmm_lasso}
\alias{plmm_lasso}
\title{Fit a penalizedLMM with lasso regularization using penalizedLMM}
\usage{
plmm_lasso(X, y, K, p1, ...)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Continuous outcome vector.}

\item{K}{Similarity matrix. For multi-chromosome analysis this may be supplied in order to perform a leave-one-chromosome-out correction. The objective here is to adjust for population stratification and unobserved confounding without rotating out the causal SNP effects.}

\item{p1}{Number of causal SNPs. Lambda will be selected such that <= p1 variables enter the model.}

\item{...}{Additional optional arguments}
}
\description{
This function allows you to fit a linear mixed model via penalized maximum likelihood with null model variance component estimation.
}
\examples{
RRM <- relatedness_mat(X = scale(admix$X))
fit <- plmm_lasso(X = admix$X, y = admix$y, K = RRM, p1 = 10)
names(fit$fit)
}
