% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_lasso.R
\name{plmm_lasso}
\alias{plmm_lasso}
\title{Fit a penalizedLMM with lasso regularization using penalizedLMM}
\usage{
plmm_lasso(X, y, p1, standardize = FALSE, X_for_K = NULL)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Continuous outcome vector.}

\item{p1}{Number of causal SNPs. Lambda will be selected such that <= p1 variables enter the model.}

\item{standardize}{Should standardization be performed within \code{glmnet()}? Defaults to FALSE.}

\item{X_for_K}{X matrix used to compute the similarity matrix, K. For multi-chromosome analysis this may be supplied in order to perform a leave-one-chromosome-out correction. The objective here is to adjust for population stratification and unobserved confounding without rotating out the causal SNP effects.}
}
\description{
This function allows you to fit a linear mixed model via penalized maximum likelihood with null model variance component estimation.
}
