% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_null.R
\name{plmm_null}
\alias{plmm_null}
\title{Estimate eta using a null Gaussian penalizedLMM model}
\usage{
plmm_null(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Continuous outcome vector.}
}
\description{
This function allows you to estimate eta (the signal to noise ratio, or narrow-sense variability) under the assumption of a null model.
}
