% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_prep.R
\name{plmm_prep}
\alias{plmm_prep}
\title{PLMM prep: a function to run checks, SVD, and rotation prior to fitting a PLMM model
This is an internal function for \code{cv_plmm}}
\usage{
plmm_prep(
  std_X,
  std_X_n,
  std_X_p,
  genomic = 1:std_X_p,
  n,
  p,
  centered_y,
  k = NULL,
  K = NULL,
  diag_K = NULL,
  eta_star = NULL,
  fbm_flag,
  penalty_factor = rep(1, ncol(std_X)),
  trace = NULL,
  ...
)
}
\arguments{
\item{std_X}{Column standardized design matrix. May include clinical covariates and other non-SNP data.}

\item{std_X_n}{The number of observations in std_X (integer)}

\item{std_X_p}{The number of features in std_X (integer)}

\item{genomic}{A numeric vector of indices indicating which columns in the standardized X are genomic covariates. Defaults to all columns.}

\item{n}{The number of instances in the \emph{original} design matrix X. This should not be altered by standardization.}

\item{p}{The number of features in the \emph{original} design matrix X, including constant features}

\item{centered_y}{Continuous outcome vector, centered.}

\item{k}{An integer specifying the number of singular values to be used in the approximation of the rotated design matrix. This argument is passed to \code{RSpectra::svds()}. Defaults to \code{min(n, p) - 1}, where n and p are the dimensions of the \emph{standardized} design matrix.}

\item{K}{Similarity matrix used to rotate the data. This should either be a known matrix that reflects the covariance of y, or an estimate (Default is \eqn{\frac{1}{p}(XX^T)}, where X is standardized). This can also be a list, with components d and u (as returned by choose_k)}

\item{diag_K}{Logical: should K be a diagonal matrix? This would reflect observations that are unrelated, or that can be treated as unrelated. Passed from \code{plmm()}.}

\item{eta_star}{Optional argument to input a specific eta term rather than estimate it from the data. If K is a known covariance matrix that is full rank, this should be 1.}

\item{fbm_flag}{Logical: is std_X an FBM type object? This is set internally by \code{plmm()}.}

\item{trace}{If set to TRUE, inform the user of progress by announcing the beginning of each step of the modeling process. Default is FALSE.}

\item{...}{Not used yet}
}
\value{
List with these components:
\itemize{
\item n: the number of rows in the original design matrix
\item p: the number of columns in the original design matrix
\item centered_y: The vector of centered outcomes
\item std_X: standardized design matrix
\item std_X_details: a list with 2 vectors:
\itemize{
\item 'center' (values used to center X)
\item 'scale' (values used to scale X)
}
\item s: vector with the eigenvalues of K
\item U: the eigenvectors of K (same as left singular values of X).
\item ns: the indices for the nonsingular values of X
\item penalty_factor: the penalty factors for the penalized non-singular values
\item feature_names: formatted column names of the design matrix
}
}
\description{
PLMM prep: a function to run checks, SVD, and rotation prior to fitting a PLMM model
This is an internal function for \code{cv_plmm}
}
\keyword{internal}
