% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm_svd.R
\name{plmm_svd}
\alias{plmm_svd}
\title{A function to implement singular value decomposition for a PLMM or LMM
This is an internal function to \code{plmm_prep()}}
\usage{
plmm_svd(K, k, trunc, trace)
}
\arguments{
\item{K}{The relatedness matrix}

\item{k}{Optional integer argument indicating the number of singular values to use in a truncated SVD. See details.}

\item{trunc}{Logical: should truncated SVD be used?}

\item{trace}{Logical: should messages be printed to console?}
}
\description{
A function to implement singular value decomposition for a PLMM or LMM
This is an internal function to \code{plmm_prep()}
}
\details{
The kind of SVD implemented here will depend on the combination of arguments supplied.
(1) if trunc = FALSE
use base::svd(K)
(2) if trunc = TRUE
use RSpectra::svds(K, k = k)
}
\keyword{internal}
