% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmmr-package.R
\docType{package}
\name{plmmr-package}
\alias{plmmr}
\alias{plmmr-package}
\title{plmmr: Penalized Linear Mixed Models For Correlated Data}
\description{
Fits penalized linear mixed models that correct for unobserved confounding factors. 'plmmr' infers and corrects for the presence of unobserved confounding effects such as population stratification and environmental heterogeneity. It then fits a linear model via penalized maximum likelihood. Originally designed for the multivariate analysis of single nucleotide polymorphisms (SNPs) measured in a genome-wide association study (GWAS), 'plmmr' eliminates the need for subpopulation-specific analyses and post-analysis p-value adjustments. Functions for the appropriate processing of 'PLINK' files are also supplied. For examples, see the package homepage. \url{https://pbreheny.github.io/plmmr/}.
}
\author{
\strong{Maintainer}: Patrick J. Breheny \email{patrick-breheny@uiowa.edu} (\href{https://orcid.org/0000-0002-0650-1119}{ORCID})

Authors:
\itemize{
  \item Tabitha K. Peter \email{tabitha-peter@uiowa.edu} (\href{https://orcid.org/0009-0005-2524-4751}{ORCID})
  \item Anna C. Reisetter \email{anna-reisetter@uiowa.edu} (\href{https://orcid.org/0000-0001-8332-4585}{ORCID})
  \item Yujing Lu
}

}
\keyword{internal}
