% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-list.R
\name{predict.list}
\alias{predict.list}
\title{Predict method for a list used in cross-validation (e.g. the \code{fit.i} within \code{cvf})}
\usage{
\method{predict}{list}(
  fit,
  newX,
  type = c("lp", "coefficients", "vars", "nvars", "blup"),
  lambda,
  idx = 1:length(fit$lambda),
  prep = NULL,
  V11 = NULL,
  V21 = NULL,
  ...
)
}
\arguments{
\item{fit}{A list with the components returned by \code{plmm_fit}}

\item{newX}{A design matrix used for computing predicted values if requested.}

\item{type}{A character argument indicating what type of prediction should be returned. Options are "lp," "coefficients," "vars," "nvars," and "blup." See details.}

\item{lambda}{A numeric vector of regularization parameter \code{lambda} values at which predictions are requested.}

\item{idx}{Vector of indices of the penalty parameter \code{lambda} at which predictions are required. By default, all indices are returned.}

\item{prep}{Optional argument. Result of the call to \code{plmm_prep} which corresponds to the \code{fit} argument. Required if \code{type == 'blup'} and object is too large to be returned in \code{fit} object.}

\item{V11}{Variance-covariance matrix of the training data. Extracted from \code{estimated_V} that is generated using all observations.}

\item{V21}{Covariance matrix between the training and the testing data. Extracted from \code{estimated_V} that is generated using all observations.}

\item{...}{Additional optional arguments}

\item{covariance}{Optional argument. $q times n$ covariance matrix between new and old observations. Required if \code{type == 'blup'}.}
}
\description{
Predict method for a list used in cross-validation (e.g. the \code{fit.i} within \code{cvf})
}
\keyword{internal}
