% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-list.R
\name{predict.list}
\alias{predict.list}
\title{Predict method for a list used in cross-validation (within \code{cvf})}
\usage{
\method{predict}{list}(
  fit,
  std_X,
  newX,
  type = c("lp", "blup"),
  idx = 1:length(fit$lambda),
  prep = NULL,
  V11 = NULL,
  V21 = NULL,
  ...
)
}
\arguments{
\item{fit}{A list with the components returned by \code{plmm_fit}.}

\item{std_X}{The standardized design matrix of training data, \emph{pre-rotation}.}

\item{newX}{A design matrix used for computing predicted values (i.e, the test data).}

\item{type}{A character argument indicating what type of prediction should be returned. Options are "lp," "coefficients," "vars," "nvars," and "blup." See details.}

\item{idx}{Vector of indices of the penalty parameter \code{lambda} at which predictions are required. By default, all indices are returned.}

\item{prep}{Optional argument. Result of the call to \code{plmm_prep} which corresponds to the \code{fit} argument. Required if \code{type == 'blup'}.}

\item{V11}{Variance-covariance matrix of the training data. Extracted from \code{estimated_V} that is generated using all observations. Required if \code{type == 'blup'}.}

\item{V21}{Covariance matrix between the training and the testing data. Extracted from \code{estimated_V} that is generated using all observations. Required if \code{type == 'blup'}.}

\item{...}{Additional optional arguments}
}
\description{
Predict method for a list used in cross-validation (within \code{cvf})
}
\details{
Define beta-hat as the coefficients estimated at the value of lambda that minimizes cross-validation error (CVE). Then options for \code{type} are as follows:
\itemize{
\item 'lp' (default): uses the linear predictor (i.e., product of new data and estimated coefficients) to predict new values of the outcome.
Note that this approach does not incorporate the correlation structure of the data.
\item 'blup' (acronym for Best Linear Unbiased Predictor): adds to the 'lp' a value that represents the estimated random effect.
This addition is a way of incorporating the estimated correlation structure of data into our prediction of the outcome.
}
}
\keyword{internal}
