% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-plmm.R
\name{predict.plmm}
\alias{predict.plmm}
\title{Predict method for "plmm" class}
\usage{
\method{predict}{plmm}(
  object,
  X,
  type = c("response", "coefficients", "vars", "nvars"),
  lambda,
  which = 1:length(object$lambda),
  ...
)
}
\arguments{
\item{object}{An object of class "plmm."}

\item{X}{Design matrix used for computing predicted values if requested.}

\item{type}{a character argument indicating what should be returned.}

\item{lambda}{A numeric vector of lambda values.}

\item{which}{Vector of lambda indices for which coefficients to return.}

\item{...}{Additional arguments.}
}
\description{
Predict method for "plmm" class
}
