% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-plmm.R
\name{predict.plmm}
\alias{predict.plmm}
\title{Predict method for plmm class}
\usage{
\method{predict}{plmm}(
  object,
  newX,
  type = c("lp", "coefficients", "vars", "nvars", "blup"),
  lambda,
  idx = 1:length(object$lambda),
  X,
  y,
  ...
)
}
\arguments{
\item{object}{An object of class \code{plmm}.}

\item{newX}{Design matrix used for computing predicted values if requested.}

\item{type}{A character argument indicating what type of prediction should be returned. Options are "lp," "coefficients," "vars," "nvars," and "blup." See details.}

\item{lambda}{A numeric vector of regularization parameter \code{lambda} values at which predictions are requested.}

\item{idx}{Vector of indices of the penalty parameter \code{lambda} at which predictions are required. By default, all indices are returned.}

\item{X}{Optional argument. Original design matrix (not including intercept column) from object. Required if \code{type == 'blup'} and object is too large to be returned in \code{plmm} object.}

\item{y}{Optional argument. Original continuous outcome vector from object. Required if \code{type == 'blup'}.}

\item{...}{Additional optional arguments}
}
\description{
Predict method for plmm class
}
\details{
Define beta-hat as the coefficients estimated at the value of lambda that minimizes cross-validation error (CVE). Then options for \code{type} are as follows:
\itemize{
\item 'response' (default): uses the product of newX and beta-hat to predict new values of the outcome. This does not incorporate the correlation structure of the data.
For the stats folks out there, this is simply the linear predictor.
\item 'blup' (acronym for Best Linear Unbiased Predictor): adds to the 'response' a value that represents the esetimated random effect. This addition is a way of incorporating
the estimated correlation structure of data into our prediction of the outcome.
\item 'coefficients': returns the estimated beta-hat
\item 'vars': returns the \emph{indicies} of variables (e.g., SNPs) with nonzero coefficients at each value of lambda. EXCLUDES intercept.
\item 'nvars': returns the \emph{number} of variables (e.g., SNPs) with nonzero coefficients at each value of lambda. EXCLUDES intercept.
}
}
\examples{
\dontrun{
# fit a model 
fit <- plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X))

# simulate new data 
newX <- sim_ps_x(n = nrow(admix$X), nJ = 4, p = ncol(admix$X),
 structureX = "independent", inbr = "heterogeneous", standardizeX = FALSE)
 
 # make predictions for all lambda values 
 pred1 <- predict(object = fit, newX = newX, type = "lp")
 
 # make predictions for a select number of lambda values 
 pred2 <- predict(object = fit, newX = newX, type = "lp", idx=98)
 }
 
 
}
