% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess PLINK files}
\usage{
preprocess(group)
}
\arguments{
\item{group}{Character argument that is the prefix of your bed/bim/fam files.}
}
\description{
This function allows you to preprocess PLINK bed/bim/fam files for use with penalizedLMM functions. Unreliable SNPs are removed and missing values are imptued using either SnpStats, or if not tagged, the HWE mean value.
}
