% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_time.R
\name{pretty_time}
\alias{pretty_time}
\title{a function to format the time}
\usage{
pretty_time()
}
\value{
A string with the formatted current date and time
}
\description{
a function to format the time
}
\keyword{internal}
