% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_plink.R
\name{process_plink}
\alias{process_plink}
\title{Preprocess PLINK files using the \code{bigsnpr} package}
\usage{
process_plink(
  data_dir,
  prefix,
  impute = TRUE,
  impute_method = "mode",
  quiet = FALSE,
  gz = FALSE,
  outfile
)
}
\arguments{
\item{data_dir}{The path to the bed/bim/fam data files}

\item{prefix}{The prefix (as a character string) of the bed/fam data files}

\item{impute}{Logical: should data be imputed? Default to TRUE.}

\item{impute_method}{If 'impute' = TRUE, this argument will specify the kind of imputation desired. This is passed directly to \code{bigsnpr::snp_fastImputeSimple()}. Defaults to 'mode'.}

\item{quiet}{Logical: should messages be printed to the console? Defaults to TRUE}

\item{gz}{Logical: are the bed/bim/fam files g-zipped? Defaults to FALSE. NOTE: if TRUE, process_plink will unzip your zipped files.}

\item{outfile}{Optional: the name (character string) of the prefix of the logfile to be written. Defaults to 'process_plink', i.e. you will get 'process_plink.log' as the outfile.
#' Note: a more nuanced 'impute' argument is still under construction. Only "simple" method is available at this time, but we hope to add "xgboost" as an option in the future.}
}
\description{
Preprocess PLINK files using the \code{bigsnpr} package
}
\examples{
\dontrun{
lite <- process_plink(data_dir = plink_example(path = "penncath_lite.bed.gz", parent = T), prefix = "penncath_lite", gz = TRUE)
mid <- process_plink(data_dir = "/Users/tabithapeter/Desktop/penalizedLMM/data-raw", prefix = "penncath_mid", gz = F, method = "mean2")
}
}
