% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_plink.R
\name{process_plink}
\alias{process_plink}
\title{Preprocess PLINK files using the \code{bigsnpr} package}
\usage{
process_plink(
  data_dir,
  prefix,
  impute = TRUE,
  impute_method = "mode",
  na_phenotype_vals = c(-9),
  id_var = "IID",
  handle_missing_phen = "prune",
  quiet = FALSE,
  gz = FALSE,
  outfile,
  overwrite = FALSE,
  add_predictor_fam = NULL,
  add_predictor_ext = NULL,
  ...
)
}
\arguments{
\item{data_dir}{The path to the bed/bim/fam data files, \emph{without} a trailing "/" (e.g., use \code{data_dir = '~/my_dir'}, \strong{not} \code{data_dir = '~/my_dir/'})}

\item{prefix}{The prefix (as a character string) of the bed/fam data files (e.g., \code{prefix = 'mydata'})}

\item{impute}{Logical: should data be imputed? Default to TRUE.}

\item{impute_method}{If 'impute' = TRUE, this argument will specify the kind of imputation desired. Options are:
\itemize{
\item mode (default): Imputes the most frequent call. See \code{bigsnpr::snp_fastImputeSimple()} for details.
\item random: Imputes sampling according to allele frequencies.
\item mean0: Imputes the rounded mean.
\item mean2: Imputes the mean rounded to 2 decimal places.
\item xgboost: Imputes using an algorithm based on local XGBoost models. See \code{bigsnpr::snp_fastImpute()} for details. Note: this can take several minutes, even for a relatively small data set.
}}

\item{na_phenotype_vals}{A vector of numeric values used to code NA values in the phenotype/outcome (this is the 'affection' column in a \code{bigSNP} object, or the last column of a \code{.fam} file). Defaults to -9 (matching PLINK conventions).}

\item{id_var}{String specifying which column of the PLINK \code{.fam} file has the unique sample identifiers. Options are "IID" (default) and "FID".}

\item{handle_missing_phen}{A string indicating how missing phenotypes should be handled:
\itemize{
\item "prune" (default): observations with missing phenotype are removed
\item "asis": leaves missing phenotypes as NA (this is fine if outcome will be supplied later from a separate file)
\item "median": impute missing phenotypes using the median (warning: this is overly simplistic in many cases).
\item "mean": impute missing phenotypes using the mean (warning: this is overly simplistic in many cases).
}}

\item{quiet}{Logical: should messages be printed to the console? Defaults to TRUE}

\item{gz}{Logical: are the bed/bim/fam files g-zipped? Defaults to FALSE. NOTE: if TRUE, process_plink will unzip your zipped files.}

\item{outfile}{Optional: the name (character string) of the prefix of the logfile to be written. Defaults to 'process_plink', i.e. you will get 'process_plink.log' as the outfile.}

\item{overwrite}{Logical: if existing \code{.bk}/\code{.rds} files exist for the specified directory/prefix, should these be overwritten? Defaults to FALSE. Set to TRUE if you want to change the imputation method you're using, etc.}

\item{add_predictor_fam}{Optional: if you want to include "sex" (the 5th column of \code{.fam} file) in the analysis, specify 'sex' here.}

\item{add_predictor_ext}{Optional: add additional covariates/predictors/features from an external file (i.e., not a PLINK file).
This argument takes one of two kinds of arguments:
(a) a \strong{named} numeric vector, where the names align with the sample IDs in the PLINK files.
The names will be used to subset and align this external covariate with the supplied PLINK data.

(b) a numeric matrix whose row names align with the sample IDs in the PLINK files.
The names will be used to subset and align this external covariate with the supplied PLINK data.}

\item{...}{Optional: additional arguments to \code{bigsnpr::snp_fastImpute()} (relevant only if impute_method = "xgboost")}
}
\value{
Nothing is returned by this function; instead, files 'prefix.rds',
'prefix.bk', and std_prefix.bk are created in the location specified by data_dir. Note that this
this function need only be run once; in subsequent data analysis/scripts,
\code{get_data()} will access the '.rds' file.
}
\description{
Preprocess PLINK files using the \code{bigsnpr} package
}
\examples{
\dontrun{

process_plink(data_dir = "inst/extdata", prefix = "admix")

process_plink(data_dir = "../temp_files",
 prefix = "penncath_lite",
  impute = T,
   quiet = F)
}
}
