% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_plink.R
\name{process_plink}
\alias{process_plink}
\title{Preprocess PLINK files using the \code{bigsnpr} package}
\usage{
process_plink(
  data_dir,
  prefix,
  impute = TRUE,
  impute_method = c("mode", "random", "mean0", "mean2", "xgboost"),
  quiet = FALSE,
  gz = FALSE,
  outfile,
  ...
)
}
\arguments{
\item{data_dir}{The path to the bed/bim/fam data files}

\item{prefix}{The prefix (as a character string) of the bed/fam data files}

\item{impute}{Logical: should data be imputed? Default to TRUE.}

\item{impute_method}{If 'impute' = TRUE, this argument will specify the kind of imputation desired. Options are:
\itemize{
\item mode (default): Imputes the most frequent call. See \code{bigsnpr::snp_fastImputeSimple()} for details.
\item random: Imputes sampling according to allele frequencies.
\item mean0: Imputes the rounded mean.
\item mean2: Imputes the mean rounded to 2 decimal places.
\item xgboost: Imputes using an algorithm based on local XGBoost models. See \code{bigsnpr::snp_fastImpute()} for details. Note: this can take several minutes, even for a relatively small data set.
}}

\item{quiet}{Logical: should messages be printed to the console? Defaults to TRUE}

\item{gz}{Logical: are the bed/bim/fam files g-zipped? Defaults to FALSE. NOTE: if TRUE, process_plink will unzip your zipped files.}

\item{outfile}{Optional: the name (character string) of the prefix of the logfile to be written. Defaults to 'process_plink', i.e. you will get 'process_plink.log' as the outfile.}

\item{...}{Optional: additional arguments to \code{bigsnpr::snp_fastImpute()} (relevant only if impute_method = "xgboost")}
}
\description{
Preprocess PLINK files using the \code{bigsnpr} package
}
\examples{
\dontrun{
process_plink(data_dir = "../temp_files",
 prefix = "penncath_lite",
  impute_method = "xgboost",
   quiet = F)
}
}
