% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_plink.R
\name{process_plink}
\alias{process_plink}
\title{Preprocess PLINK files}
\usage{
process_plink(
  prefix,
  dataDir,
  sexcheck = FALSE,
  na.strings = "-9",
  impute = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{prefix}{Character argument that is the prefix of your bed/bim/fam files.}

\item{dataDir}{Directory where plink files (and .sexcheck files are located if \code{sexcheck = TRUE}) are located}

\item{sexcheck}{Logical flag for whether or not PLINK sexcheck files should be incorporated. Defaults to FALSE. If TRUE, sexcheck files must be of the form "prefix.sexcheck"}

\item{na.strings}{For \code{snpStats}. Strings in .bam and .fam files to be recoded as NA. Defaults to "-9"}

\item{impute}{Logical flag for whether imputation should be performed. Defaults to TRUE since plmm cannot handle missing values.}

\item{quiet}{Logical flag: should progress messages be printed? Defaults to FALSE.}
}
\value{
A three element list object:
\itemize{
\item \code{genotypes} The filtered and imputed genotypes in a snpMatrix object with subjects in rows and SNPs in columns. This is a numeric matrix.
\item \code{map} A matrix of SNP data.
\item \code{fam} A matrix of subject data.
}
}
\description{
This function allows you to preprocess PLINK bed/bim/fam files for use with \code{penalizedLMM} functions. Unreliable SNPs are removed and missing values are imptued using either \code{snpStats}, or if not tagged, the HWE mean value.
}
\examples{
test <- process_plink(prefix = "cad_lite", dataDir = plink_example(path="cad_lite.fam", parent=T))
\dontrun{
test2 <- process_plink(prefix = "cad_mid", dataDir = plink_example(path="cad_mid.fam", parent=T))
}


}
