% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_plink_2.R
\name{process_plink_2}
\alias{process_plink_2}
\title{An updated function for processing PLINK files}
\usage{
process_plink_2(
  data_dir,
  prefix,
  rds = FALSE,
  impute = "simple",
  quiet = FALSE,
  ...
)
}
\arguments{
\item{data_dir}{The path to the bed/bim/fam data files}

\item{prefix}{The prefix (as a character string) of the bed/fam data files}

\item{rds}{Logical: does an rds file for these data files already exist in \code{data_dir}? Defaults to FALSE}

\item{impute}{Character string indicating type of imputation to use. "simple" will use \code{bigsnpr::snp_fastImputeSimple}, and "xgboost" will use \code{bigsnpr::snp_fastImpute}. Defaults to "simple."
Note: 'impute' argument is still under construction. Only "simple" method is available at this time.}

\item{quiet}{Logical: should messages be printed to the console? Defaults to TRUE}

\item{...}{Other arguments to bigsnpr::snp_fastImpute or bigsnpr::snp_fastImputeSimple (depending on choice of \code{impute})}
}
\description{
An updated function for processing PLINK files
}
\examples{

\dontrun{
cad_mid <- process_plink_2(data_dir = plink_example(path = "cad_mid.bed", parent = T), prefix = "cad_mid", rds = T, impute = "simple", method = "mean0")
cad_lite <- process_plink_2(data_dir = plink_example(path = "cad_lite.bed", parent = T), prefix = "cad_lite", rds = F, impute = "xgboost")
}
}
