% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness_mat.R
\name{relatedness_mat}
\alias{relatedness_mat}
\title{Calculate a relatedness matrix}
\usage{
relatedness_mat(X, std = TRUE, fbm = FALSE, ns = NULL, ...)
}
\arguments{
\item{X}{A numeric matrix of genotypes (from \emph{fully-imputed} data)}

\item{std}{Logical: should X be standardized? If you set this to FALSE, you should know exactly why that is appropriate... standardization is a best practice, and this will impact results.}

\item{fbm}{Logical: is X stored as an FBM? Defaults to FALSE}

\item{ns}{Optional vector of values indicating the indices of nonsingular features}

\item{...}{Other optional arguments to \code{bigstatsr::bigapply()} (like \code{ncores = ...})}
}
\description{
This function allows you to generate an n by n genetic relatedness matrix. If a numeric matrix is supplied, the RRM (Hayes, 2009) is used
and is computed XX'/p, where X is standardized.
}
\examples{
RRM <- relatedness_mat(X = admix$X)
}
