% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{a function to rotate data (or 'precondition' data) onto the transformed scale}
\usage{
rotate(s, U, eta, y = NULL, X = NULL)
}
\arguments{
\item{s}{vector of eigenvalues of X}

\item{U}{left singular vectors of X}

\item{eta}{estimated variance parameter}

\item{y}{vector of outcomes}

\item{X}{design matrix}
}
\description{
a function to rotate data (or 'precondition' data) onto the transformed scale
}
\keyword{internal}
