% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_data.R
\name{rotate_data}
\alias{rotate_data}
\title{Compute transformed X and y for subsequent plmm rescaling and fitting}
\usage{
rotate_data(X, y, K = NULL, eta_star)
}
\arguments{
\item{X}{Design matrix. May include clinical covariates and other non-SNP data.}

\item{y}{Continuous outcome vector.}

\item{K}{Estimated or known similarity matrix. By default, K is the realized relationship matrix, \eqn{\frac{1}{p}XX^T}, where \eqn{p} is the number of columns in X}

\item{eta_star}{Optional arg to input a specific eta term rather than estimate it from the data. If v is a known matrix, this should be 1.}
}
\description{
This function allows you to fit a linear mixed model via non-convex penalized maximum likelihood.
}
\examples{
std_X <- scale(admix$X)
K <- relatedness_mat(std_X)
rotated_dat <- rotate_data(std_X, admix$y, K)
}
