% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_filebacked.R
\name{rotate_filebacked}
\alias{rotate_filebacked}
\title{A function to rotate filebacked data}
\usage{
rotate_filebacked(prep, ...)
}
\value{
a list with 4 items:
\itemize{
\item rot_X" \code{X} on the rotated scale, without re-standardizing (an FBM)
\item roty_y: \code{y} on the rotated scale (a numeric vector)
\item stdrot_X: \code{X} on the rotated scale, re-standardized (an FBM)
\item stdrot_X_scale: numeric vector of values used to standardize \code{rot_X}
}
}
\description{
A function to rotate filebacked data
}
\keyword{internal}
