% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{SCAD}
\alias{SCAD}
\title{helper function to implement SCAD penalty}
\usage{
SCAD(z, l1, l2, gamma, v)
}
\arguments{
\item{z}{solution over active set at each feature}

\item{l1}{upper bound}

\item{l2}{lower bound}

\item{gamma}{The tuning parameter of the SCAD penalty}

\item{v}{the 'xtx' term}
}
\description{
helper function to implement SCAD penalty
}
\keyword{internal}
