% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_varp.R
\name{scale_varp}
\alias{scale_varp}
\title{Calculate scale by the population standard deviation, without centering}
\usage{
scale_varp(X)
}
\arguments{
\item{X}{numeric matrix}
}
\description{
This function allows you to scale vectors of a matrix by their population standard deviation without centering; we assume our sample is the population.
}
\examples{
M <- matrix(rnorm(25), 5, 5)
head(M)
M_scaled <- scale_varp(M)
head(M_scaled)

X_scaled <- scale_varp(admix$X)
admix$X[1:5, 1:7]; X_scaled[1:5, 1:7]
}
