% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_lambda.R
\name{setup_lambda}
\alias{setup_lambda}
\title{Compute sequence of lambda values}
\usage{
setup_lambda(X, y, alpha, lambda.min, nlambda, penalty.factor)
}
\arguments{
\item{X}{Design matrix which \emph{includes} the intercept column if present. May include clinical covariates and other non-SNP data.}

\item{y}{Continuous outcome vector.}

\item{alpha}{Tuning parameter for the Mnet estimator which controls the relative contributions from the MCP/SCAD penalty and the ridge, or L2 penalty. alpha=1 is equivalent to MCP/SCAD penalty, while alpha=0 would be equivalent to ridge regression. However, alpha=0 is not supported; alpha may be arbitrarily small, but not exactly 0.}

\item{lambda.min}{The smallest value for lambda, as a fraction of lambda.max. Default is .001 if the number of observations is larger than the number of covariates and .05 otherwise.}

\item{nlambda}{The smallest value for lambda, as a fraction of lambda.max. Default is .001 if the number of observations is larger than the number of covariates and .05 otherwise.}

\item{penalty.factor}{A multiplicative factor for the penalty applied to each coefficient. If supplied, penalty.factor must be a numeric vector of length equal to the number of columns of X. The purpose of penalty.factor is to apply differential penalization if some coefficients are thought to be more likely than others to be in the model. In particular, penalty.factor can be 0, in which case the coefficient is always in the model without shrinkage.}
}
\description{
This function allows you compute a sequence of lambda values for plmm models.
}
