% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_fbm.R
\name{standardize_fbm}
\alias{standardize_fbm}
\title{A helper function to standardize an FBM}
\usage{
standardize_fbm(
  obj,
  prefix,
  rds_dir,
  non_gen,
  complete_phen,
  id_var,
  outfile,
  quiet
)
}
\arguments{
\item{obj}{A \code{bigSNP} object}

\item{prefix}{The prefix (as a character string) of the bed/fam data files (e.g., \code{prefix = 'mydata'})}

\item{rds_dir}{The path to the directory in which you want to create the new '.rds' and '.bk' files. Defaults to \code{data_dir}}

\item{non_gen}{An integer vector that ranges from 1 to the number of added predictors. Example: if 2 predictors are added, non_gen = 1:2.
Note: this is typically passed from the result of \code{add_predictors()}}

\item{complete_phen}{Numeric vector with indicies marking the rows of the original data which have a non-missing entry in the 6th column of the \code{.fam} file}

\item{id_var}{String specifying which column of the PLINK \code{.fam} file has the unique sample identifiers. Options are "IID" (default) and "FID".}

\item{outfile}{Optional: the name (character string) of the prefix of the logfile to be written. Defaults to 'process_plink', i.e. you will get 'process_plink.log' as the outfile.}

\item{quiet}{Logical: should messages be printed to the console? Defaults to TRUE}
}
\value{
A list with a new component of \code{obj} called 'std_X' - this is an FBM with column-standardized data.
List also includes several other indices/meta-data on the standardized matrix
}
\description{
A helper function to standardize an FBM
}
\keyword{internal}
