% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_matrix.R
\name{standardize_matrix}
\alias{standardize_matrix}
\title{A helper function to standardize matrices}
\usage{
standardize_matrix(X, penalty_factor)
}
\arguments{
\item{X}{a matrix}

\item{penalty_factor}{a vector of 0s and 1s, passed from \code{plmm()}}
}
\value{
a list with the standardized matrix and its details
}
\description{
A helper function to standardize matrices
}
\keyword{internal}
