% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-plmm.R
\name{summary.plmm}
\alias{summary.plmm}
\title{A summary method for the plmm class}
\usage{
\method{summary}{plmm}(object, quiet = FALSE)
}
\arguments{
\item{object}{An object of class \code{plmm}}

\item{quiet}{A logical indicating whether the console output is desired. Defaults to FALSE.}
}
\value{
A list containing:
}
\description{
A summary method for the plmm class
}
\examples{
fit <- plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X))
example <- summary.plmm(fit)
#TODO: Sept. 14, 2022 - need a more sophisticated way to select the 'best' lambda value 
}
