% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eta.R
\name{test_eta_estimation}
\alias{test_eta_estimation}
\title{a function to write tests of estimate_eta}
\usage{
test_eta_estimation(sig_s, sig_eps, K, ...)
}
\arguments{
\item{sig_s}{Variance attributable to structure}

\item{sig_eps}{Variance of random error}

\item{K}{Matrix to use as relatedness matrix.}

\item{...}{Additional args to pass into \code{estimate_eta()}}
}
\description{
a function to write tests of estimate_eta
}
\keyword{internal}
