% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstandardize.R
\name{unstandardize}
\alias{unstandardize}
\title{Unstandardize coefficient values}
\usage{
unstandardize(b, X, XX, intercept = TRUE)
}
\arguments{
\item{b}{p x nlambda matrix of standardized coefficient path values.}

\item{X}{Original, non-standardized design matrix without an intercept column. Necessary for properly defining the dimensions of beta in cases where singular columns are present.}

\item{XX}{Standardized design matrix. Must include the attributes 'center', 'scale', and 'nonsingular' corresponding to non-intercept variables.}

\item{intercept}{Logical flag for whether an intercept row is included in b, that is, a coefficient which accounts for the mean of the original y value on its rotated scale.}
}
\description{
This function allows you to unstandardize coefficient values based on attributes of X.
}
