% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstandardize.R
\name{unstandardize}
\alias{unstandardize}
\title{Unstandardize coefficient values}
\usage{
unstandardize(b, X, std_X, intercept = TRUE)
}
\arguments{
\item{b}{p x nlambda matrix of standardized coefficient path values.}

\item{X}{Original, non-standardized design matrix without an intercept column. Necessary for properly defining the dimensions of beta in cases where singular columns are present.}

\item{std_X}{Standardized design matrix. Must include the attributes 'center', 'scale', and 'nonsingular' corresponding to non-intercept variables.}

\item{intercept}{Logical flag for whether an intercept row is included in b, that is, a coefficient which accounts for the mean of the original y value on its rotated scale.}
}
\description{
This function allows you to unstandardize coefficient values based on attributes of X.
}
\examples{
admix$K <- (admix$X\%*\%t(admix$X))/ncol(admix$X) # create an estimated covariance matrix 
fit <- plmm(X = admix$X, y = admix$y, K = admix$K)
# TODO: this throws an error: new_betas <- unstandardize(b = fit$beta, X = admix$X, std_X = scale(admix$X))
}
