% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untransform.R
\name{untransform}
\alias{untransform}
\title{Untransform coefficient values back to the original scale}
\usage{
untransform(untransformed_b1, std_X_details, fbm_flag)
}
\arguments{
\item{untransformed_b1}{The estimated coefficients on the standardized scale}

\item{std_X_details}{A list with 3 elements describing the standardized design matrix BEFORE rotation; this should have elements 'scale', 'center', and 'ns'}

\item{fbm_flag}{Logical: is the corresponding design matrix filebacked?}
}
\description{
This function unwinds the initial standardization of the data to obtain
coefficient values on their original scale. It is called by plmm_format().
}
\keyword{internal}
