% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untransform.R
\name{untransform}
\alias{untransform}
\title{Untransform coefficient values back to the original scale}
\usage{
untransform(res_b, ns, X, std_X, SUX, std_SUX)
}
\arguments{
\item{res_b}{The values returned in the 'beta' argument of the ncvfit() object}

\item{ns}{The indices of the non-singular columns of the ORIGINAL design matrix}

\item{X}{The original design matrix (has no intercept)}

\item{std_X}{The standardized design matrix BEFORE rotation; this should have attributes 'scale', 'center', and 'nonsingular'}

\item{SUX}{The rotated design matrix (has the intercept)}

\item{std_SUX}{The standardized design matrix AFTER rotation; this should have attribute 'scale'}
}
\description{
This function unwinds both steps of the standardization process to obtain
coefficient values on their original scale. It is called by plmm().
}
