% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untransform.R
\name{untransform}
\alias{untransform}
\title{Untransform coefficient values back to the original scale}
\usage{
untransform(std_scale_beta, p, std_X_details, fbm_flag, non_genomic)
}
\arguments{
\item{std_scale_beta}{The estimated coefficients on the standardized scale}

\item{p}{The number of columns in the original design matrix}

\item{std_X_details}{A list with 3 elements describing the standardized design matrix BEFORE rotation; this should have elements 'scale', 'center', and 'ns'}

\item{fbm_flag}{Logical: is the corresponding design matrix filebacked?}

\item{non_genomic}{Optional vector specifying which columns of the design matrix represent features that are \emph{not} genomic, as these features are excluded from the empirical estimation of genomic relatedness.
For cases where X is a filepath to an object created by \code{process_plink()}, this is handled automatically via the arguments to \code{process_plink()}.
For all other cases, 'non_genomic' defaults to NULL (meaning \code{plmm()} will assume that all columns of \code{X} represent genomic features).}
}
\description{
This function unwinds the initial standardization of the data to obtain
coefficient values on their original scale. It is called by plmm_format().
}
\keyword{internal}
