% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_data.R
\name{unzip_example_data}
\alias{unzip_example_data}
\title{a function to unzip the .gz files that ship with the package
\emph{Note}: for Linux/Unix and MacOS only}
\usage{
unzip_example_data(outdir)
}
\arguments{
\item{outdir}{The filepath to the directory to which the .gz files should be written}
}
\description{
a function to unzip the .gz files that ship with the package
\emph{Note}: for Linux/Unix and MacOS only
}
